feat(wallet): unify restore path, multisig 25th-word support, shared verify/protect sections
Restore/import is now ONE page. RestoreWalletFlowComponent
(features/wallet-setup, single-page mobile shape): shared name section,
mnemonic entry, BIP39 25th word, at-rest secret, commit branching Core
RPC (rescan+restore branch report) vs BDK (Electrum branch probe) on
NodeService.isRemote. Desktop "import wallet" and mobile restore are
thin hosts; desktop gains the richer success card (branch label,
legacy-branch note, scan-again reprobe) and the prefilled name.

Multisig wizard:
- BIP39 25th-word support end-to-end - the derivation layer already
  accepted a passphrase (mnemonicToSeedSync), the wizard now collects it
  (shared section, create/restore modes) and passes it into
  deriveMultisigKey + generateMultisigDescriptors
- shared name section / verify step / at-rest section replace the
  hand-rolled near-duplicates
- save-key/save-descriptor buttons routed through downloadTextFile (the
  detached-anchor download is a silent no-op in the Tauri webview)

Shared extractions (dogfooded by the create flow):
- VerifyWordsComponent - the 3-random-words check with autocomplete,
  Enter-to-advance, per-field errors; rolls fresh indices per mount
- AtRestPassphraseSectionComponent - lean optional at-rest secret with
  confirm + forget-warning once typed, per-flow hint key

Watch-only adopts the shared name section too - that was the fourth
hand-rolled copy of the name field, all four carrying the same
mat-error-never-renders bug the shared section fixes.

i18n: 17 more orphaned keys deleted across all 26 locales (legacy
encryption set, import/restore duplicates, mwallet_* titles).
