+++ web-wallet/src/app/shared/components/index.ts
@@ -41,4 +41,6 @@ export { StepHeaderComponent } from './step-header/step-header.component';
 export { MnemonicDisplayComponent } from './mnemonic-display/mnemonic-display.component';
 export { MnemonicEntryComponent } from './mnemonic-entry/mnemonic-entry.component';
 export { Bip39PassphraseSectionComponent } from './bip39-passphrase-section/bip39-passphrase-section.component';
+export { VerifyWordsComponent } from './verify-words/verify-words.component';
+export { AtRestPassphraseSectionComponent } from './at-rest-passphrase-section/at-rest-passphrase-section.component';
 export type { MnemonicEntryState } from './mnemonic-entry/mnemonic-entry.component';
