+++ web-wallet/src/assets/locales/bg.json
@@ -140,7 +140,6 @@
   "confirm_send": "Потвърди и изпрати",
   "confirm_send_message": "Моля, прегледайте детайлите на транзакцията преди изпращане.",
   "confirm_transaction": "Потвърдете транзакцията",
-  "confirm_wallet_password": "Потвърдете паролата",
   "confirmations": "Потвърждения",
   "confirmations_short": "потв.",
   "confirmed": "Потвърдено",
@@ -224,7 +223,6 @@
   "edit_contact": "Редактиране на контакт",
   "effective_at_block": "В сила на блок",
   "effective_height": "Действаща височина",
-  "encrypt_wallet": "Шифровай портфейла с парола",
   "enter_forging_address": "Въведете ковашки или пул адрес",
   "error": "Грешка",
   "error_checking_status": "Грешка при проверка на статус",
@@ -269,10 +267,8 @@
   "import": "Импортирай",
   "import_mnemonic_description": "Въведете вашата фраза за възстановяване от 12 или 24 думи, за да възстановите портфейла си.",
   "import_to_wallet": "Импортирай в портфейла",
-  "import_wallet": "Импортирай портфейл",
   "import_wif_description": "Импортирайте WIF частен ключ като wpkh (Native SegWit) дескриптор в портфейла",
   "import_wif_wpkh": "Импортирай WIF ключ (wpkh)",
-  "importing": "Импортиране",
   "inbound": "Входящи",
   "incoming_payment": "Входящо плащане",
   "incorrect_word": "Неправилна дума",
@@ -541,7 +537,6 @@
   "passphrase": "Фраза за възстановяване",
   "passphrase_mismatch": "Паролите не съвпадат",
   "password": "Парола",
-  "password_mismatch": "Паролите не съвпадат",
   "payment_confirmed": "Плащане потвърдено",
   "payment_uri": "URI за плащане",
   "payment_uri_copied": "URI за плащане е копиран в клипборда",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "стар десктоп",
   "restore_branch_pocx": "мобилен",
   "restore_history_found": "Открити са средства на: {branches}",
-  "restore_no_history": "Не е открита предишна история за тази фраза — портфейлът започва начисто.",
   "reset": "Нулиране",
   "reset_mining_config": "Нулиране на конфигурацията за добив",
   "reset_mining_config_confirm": "Това ще изчисти всички настройки за добив и ще се върне към съветника за настройка. Плот файловете няма да бъдат изтрити. Сигурни ли сте?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Можете да шифровате портфейла си с парола по избор. Това защитава файла на портфейла и изисква паролата за достъп до средствата ви.",
   "wallet_encryption_title": "Шифроване на портфейл",
   "wallet_encryption_warning": "Ако забравите тази парола, все още можете да възстановите портфейла си с 24-думната фраза. Въпреки това ще трябва да създадете нов портфейл.",
-  "wallet_imported_success": "Портфейлът '{name}' е импортиран успешно",
   "wallet_lock_status_all_locked": "Всички криптирани портфейли заключени",
   "wallet_lock_status_some_unlocked": "Отключени {unlocked} от {total} криптирани портфейли",
   "wallet_locked_message": "Този портфейл е заключен. Моля, въведете паролата за временно отключване.",
   "wallet_name": "Име на портфейл",
-  "wallet_name_hint": "Изберете име за идентифициране на този портфейл",
   "wallet_not_encrypted": "Не е шифрован",
-  "wallet_password": "Парола на портфейл",
   "wallet_status_notifications": "Известия за статус на портфейл",
   "wallets": "Портфейл(и)",
   "warn": "Предупр.",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "синхронизирано преди {seconds} с",
   "wallet_name_invalid_local": "Използвайте 1–32 букви, цифри, „-“ или „_“",
   "wallet_name_hint_local": "Само букви, цифри, „-“ и „_“",
-  "wallet_encryption_info_local": "По желание шифровайте семето на портфейла на диска с парола. Ще ви е необходима, за да отключите портфейла за изпращане. Само фразата за възстановяване винаги възстановява средствата ви.",
-  "restore_probe_info": "Историята на вашето семе се проверява през Electrum по време на импортирането — клоновете със съществуващи средства се откриват автоматично.",
   "psbt_join_unavailable_remote": "Обединяването не е налично в отдалечен режим",
   "feature_unavailable_remote": "Тази функция не е налична в отдалечен режим (Electrum)",
   "setup_solo_not_available_remote": "Соло копаенето изисква локален възел — не е налично в отдалечен режим (Electrum)",
