+++ web-wallet/src/assets/locales/ca.json
@@ -140,7 +140,6 @@
   "confirm_send": "Confirmar i enviar",
   "confirm_send_message": "Si us plau, reviseu els detalls de la transacció abans d'enviar.",
   "confirm_transaction": "Confirmar transacció",
-  "confirm_wallet_password": "Confirmar contrasenya",
   "confirmations": "Confirmacions",
   "confirmations_short": "conf",
   "confirmed": "Confirmat",
@@ -224,7 +223,6 @@
   "edit_contact": "Edita el contacte",
   "effective_at_block": "Efectiu al bloc",
   "effective_height": "Alçada efectiva",
-  "encrypt_wallet": "Xifrar cartera amb contrasenya",
   "enter_forging_address": "Introduïu l'adreça de forja o de pool",
   "error": "Error",
   "error_checking_status": "Error en verificar l'estat",
@@ -269,10 +267,8 @@
   "import": "Importa",
   "import_mnemonic_description": "Introduïu la vostra frase de recuperació de 12 o 24 paraules per restaurar la vostra cartera.",
   "import_to_wallet": "Importar a la cartera",
-  "import_wallet": "Importar cartera",
   "import_wif_description": "Importar una clau privada WIF com a descriptor wpkh (Native SegWit) a la cartera",
   "import_wif_wpkh": "Importar clau WIF (wpkh)",
-  "importing": "Important",
   "inbound": "Entrant",
   "incoming_payment": "Pagament entrant",
   "incorrect_word": "Paraula incorrecta",
@@ -541,7 +537,6 @@
   "passphrase": "Frase de recuperació",
   "passphrase_mismatch": "Les frases de contrasenya no coincideixen",
   "password": "Contrasenya",
-  "password_mismatch": "Les contrasenyes no coincideixen",
   "payment_confirmed": "Pagament confirmat",
   "payment_uri": "URI de pagament",
   "payment_uri_copied": "URI de pagament copiada al porta-retalls",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "escriptori antic",
   "restore_branch_pocx": "mòbil",
   "restore_history_found": "S'han trobat fons a: {branches}",
-  "restore_no_history": "No s'ha trobat historial previ per a aquesta frase — la cartera comença de zero.",
   "reset": "Restableix",
   "reset_mining_config": "Restablir configuració de mineria",
   "reset_mining_config_confirm": "Això esborrarà tota la configuració de mineria i tornarà a l'assistent de configuració. Els fitxers de plot no s'eliminaran. Esteu segur?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Opcionalment podeu xifrar la vostra cartera amb una contrasenya. Això protegeix el fitxer de la vostra cartera i requereix la contrasenya per accedir als vostres fons.",
   "wallet_encryption_title": "Xifratge de cartera",
   "wallet_encryption_warning": "Si oblideu aquesta contrasenya, encara podreu recuperar la vostra cartera usant la vostra frase llavor de 24 paraules. No obstant això, haureu de crear una nova cartera.",
-  "wallet_imported_success": "Cartera '{name}' importada amb èxit",
   "wallet_lock_status_all_locked": "Totes les carteres xifrades bloquejades",
   "wallet_lock_status_some_unlocked": "{unlocked} de {total} carteres xifrades desbloquejades",
   "wallet_locked_message": "Aquesta cartera està bloquejada. Si us plau, introduïu la frase de contrasenya per desbloquejar-la temporalment.",
   "wallet_name": "Nom de la cartera",
-  "wallet_name_hint": "Trieu un nom per identificar aquesta cartera",
   "wallet_not_encrypted": "Sense xifrar",
-  "wallet_password": "Contrasenya de la cartera",
   "wallet_status_notifications": "Notificacions d'estat de la cartera",
   "wallets": "Cartera(es)",
   "warn": "Advert.",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "sincronitzat fa {seconds}s",
   "wallet_name_invalid_local": "Fes servir d'1 a 32 lletres, dígits, '-' o '_'",
   "wallet_name_hint_local": "Només lletres, dígits, '-' i '_'",
-  "wallet_encryption_info_local": "Opcionalment, xifra la llavor del moneder al disc amb una contrasenya. La necessitaràs per desbloquejar el moneder en enviar. La frase de recuperació per si sola sempre restaura els teus fons.",
-  "restore_probe_info": "L'historial de la teva llavor es comprova mitjançant Electrum durant la importació — les branques amb fons existents es detecten automàticament.",
   "psbt_join_unavailable_remote": "Combinar no està disponible en mode remot",
   "feature_unavailable_remote": "Aquesta funció no està disponible en mode remot (Electrum)",
   "setup_solo_not_available_remote": "La mineria en solitari necessita un node local — no disponible en mode remot (Electrum)",
