+++ web-wallet/src/assets/locales/cs.json
@@ -140,7 +140,6 @@
   "confirm_send": "Potvrdit a odeslat",
   "confirm_send_message": "Zkontrolujte podrobnosti transakce pred odeslanim.",
   "confirm_transaction": "Potvrdit transakci",
-  "confirm_wallet_password": "Potvrdit heslo",
   "confirmations": "Potvrzeni",
   "confirmations_short": "potv.",
   "confirmed": "Potvrzeno",
@@ -224,7 +223,6 @@
   "edit_contact": "Upravit kontakt",
   "effective_at_block": "Ucinne od bloku",
   "effective_height": "Účinná výška",
-  "encrypt_wallet": "Zasifrovat penezenku heslem",
   "enter_forging_address": "Zadejte adresu kovani nebo poolu",
   "error": "Chyba",
   "error_checking_status": "Chyba pri kontrole stavu",
@@ -269,10 +267,8 @@
   "import": "Importovat",
   "import_mnemonic_description": "Zadejte svou 12 nebo 24 slovni obnovovaci frazi pro obnoveni penezenky.",
   "import_to_wallet": "Importovat do penezenky",
-  "import_wallet": "Importovat penezenku",
   "import_wif_description": "Importovat soukromy klic WIF jako wpkh (Native SegWit) deskriptor do penezenky",
   "import_wif_wpkh": "Importovat WIF klic (wpkh)",
-  "importing": "Importovani",
   "inbound": "Prichozi",
   "incoming_payment": "Prichozi platba",
   "incorrect_word": "Nespravne slovo",
@@ -541,7 +537,6 @@
   "passphrase": "Obnovovaci fraze",
   "passphrase_mismatch": "Hesla se neshoduji",
   "password": "Heslo",
-  "password_mismatch": "Hesla se neshoduji",
   "payment_confirmed": "Platba potvrzena",
   "payment_uri": "URI platby",
   "payment_uri_copied": "URI platby zkopirovano do schranky",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "starší desktop",
   "restore_branch_pocx": "mobilní",
   "restore_history_found": "Prostředky nalezeny na: {branches}",
-  "restore_no_history": "Pro tuto frázi nebyla nalezena žádná dřívější historie — peněženka začíná od nuly.",
   "reset": "Resetovat",
   "reset_mining_config": "Resetovat konfiguraci tezby",
   "reset_mining_config_confirm": "Tim se vymaze vsechna nastaveni tezby a vratite se do pruvodce nastavenim. Plot soubory nebudou smazany. Opravdu chcete pokracovat?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Muzete volitelne zasifrovat svou penezenku heslem. To chrani soubor penezenky a vyzaduje heslo pro pristup k prostredkum.",
   "wallet_encryption_title": "Sifrovani penezenky",
   "wallet_encryption_warning": "Pokud toto heslo zapomenete, muzete stale obnovit penezenku pomoci sve 24slovni seed fraze. Budete vsak muset vytvorit novou penezenku.",
-  "wallet_imported_success": "Penezenka '{name}' uspesne importovana",
   "wallet_lock_status_all_locked": "Vsechny sifrovane penezenky uzamceny",
   "wallet_lock_status_some_unlocked": "Odemceno {unlocked} z {total} sifrovanych penezenek",
   "wallet_locked_message": "Tato penezenka je uzamcena. Zadejte heslo pro docasne odemceni.",
   "wallet_name": "Nazev penezenky",
-  "wallet_name_hint": "Zvolte nazev pro identifikaci teto penezenky",
   "wallet_not_encrypted": "Nezasifrovano",
-  "wallet_password": "Heslo penezenky",
   "wallet_status_notifications": "Oznameni o stavu penezenky",
   "wallets": "Penezenka(y)",
   "warn": "Varování",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "synchronizováno před {seconds} s",
   "wallet_name_invalid_local": "Použijte 1–32 písmen, číslic, „-“ nebo „_“",
   "wallet_name_hint_local": "Pouze písmena, číslice, „-“ a „_“",
-  "wallet_encryption_info_local": "Volitelně zašifrujte semeno peněženky na disku heslem. Budete ho potřebovat k odemknutí peněženky pro odesílání. Samotná obnovovací fráze vždy obnoví vaše prostředky.",
-  "restore_probe_info": "Historie vašeho semene se během importu prověřuje přes Electrum — větve s existujícími prostředky jsou detekovány automaticky.",
   "psbt_join_unavailable_remote": "Sloučení není ve vzdáleném režimu k dispozici",
   "feature_unavailable_remote": "Tato funkce není ve vzdáleném režimu (Electrum) k dispozici",
   "setup_solo_not_available_remote": "Sólo těžba vyžaduje lokální uzel — není k dispozici ve vzdáleném režimu (Electrum)",
