+++ web-wallet/src/assets/locales/de-de.json
@@ -140,7 +140,6 @@
   "confirm_send": "Bestaetigen & Senden",
   "confirm_send_message": "Bitte ueberpruefe die Transaktionsdetails vor dem Senden.",
   "confirm_transaction": "Transaktion bestaetigen",
-  "confirm_wallet_password": "Passwort bestaetigen",
   "confirmations": "Bestaetigungen",
   "confirmations_short": "Best.",
   "confirmed": "Bestaetigt",
@@ -224,7 +223,6 @@
   "edit_contact": "Kontakt bearbeiten",
   "effective_at_block": "Wirksam ab Block",
   "effective_height": "Wirksamkeitshoehe",
-  "encrypt_wallet": "Wallet mit Passwort verschluesseln",
   "enter_forging_address": "Forging- oder Pool-Adresse eingeben",
   "error": "Fehler",
   "error_checking_status": "Fehler bei der Statuspruefung",
@@ -269,10 +267,8 @@
   "import": "Importieren",
   "import_mnemonic_description": "Gib deine 12 oder 24 Woerter Wiederherstellungsphrase ein, um dein Wallet wiederherzustellen.",
   "import_to_wallet": "In Wallet importieren",
-  "import_wallet": "Wallet importieren",
   "import_wif_description": "Importiere einen privaten WIF-Schluessel als wpkh (Native SegWit) Deskriptor in das Wallet",
   "import_wif_wpkh": "WIF-Schluessel importieren (wpkh)",
-  "importing": "Importiere",
   "inbound": "Eingehend",
   "incoming_payment": "Eingehende Zahlung",
   "incorrect_word": "Falsches Wort",
@@ -541,7 +537,6 @@
   "passphrase": "Wiederherstellungsphrase",
   "passphrase_mismatch": "Passphrasen stimmen nicht ueberein",
   "password": "Passwort",
-  "password_mismatch": "Passwoerter stimmen nicht ueberein",
   "payment_confirmed": "Zahlung bestaetigt",
   "payment_uri": "Zahlungs-URI",
   "payment_uri_copied": "Zahlungs-URI in Zwischenablage kopiert",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "alter Desktop",
   "restore_branch_pocx": "Mobil",
   "restore_history_found": "Guthaben gefunden auf: {branches}",
-  "restore_no_history": "Für diese Phrase wurde keine frühere Historie gefunden — die Wallet startet neu.",
   "reset": "Zuruecksetzen",
   "reset_mining_config": "Mining-Konfiguration zuruecksetzen",
   "reset_mining_config_confirm": "Dies loescht alle Mining-Einstellungen und kehrt zum Setup-Assistenten zurueck. Plot-Dateien werden nicht geloescht. Bist du sicher?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Du kannst optional dein Wallet mit einem Passwort verschluesseln. Dies schuetzt deine Wallet-Datei und erfordert das Passwort, um auf deine Gelder zuzugreifen.",
   "wallet_encryption_title": "Wallet-Verschluesselung",
   "wallet_encryption_warning": "Wenn du dieses Passwort vergisst, kannst du dein Wallet immer noch mit deiner 24-Woerter-Seed-Phrase wiederherstellen. Du musst jedoch ein neues Wallet erstellen.",
-  "wallet_imported_success": "Wallet '{name}' erfolgreich importiert",
   "wallet_lock_status_all_locked": "Alle verschluesselten Wallets gesperrt",
   "wallet_lock_status_some_unlocked": "{unlocked} von {total} verschluesselten Wallets entsperrt",
   "wallet_locked_message": "Dieses Wallet ist gesperrt. Bitte gib das Passwort ein, um es voruebergehend zu entsperren.",
   "wallet_name": "Wallet-Name",
-  "wallet_name_hint": "Waehle einen Namen, um dieses Wallet zu identifizieren",
   "wallet_not_encrypted": "Nicht verschluesselt",
-  "wallet_password": "Wallet-Passwort",
   "wallet_status_notifications": "Wallet-Statusbenachrichtigungen",
   "wallets": "Wallet(s)",
   "warn": "Warnung",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "vor {seconds}s synchronisiert",
   "wallet_name_invalid_local": "1–32 Buchstaben, Ziffern, '-' oder '_' verwenden",
   "wallet_name_hint_local": "Nur Buchstaben, Ziffern, '-' und '_'",
-  "wallet_encryption_info_local": "Optional die Wallet-Seed auf der Festplatte mit einem Passwort verschlüsseln. Du brauchst es, um die Wallet zum Senden zu entsperren. Die Wiederherstellungsphrase allein stellt deine Guthaben immer wieder her.",
-  "restore_probe_info": "Der Verlauf deiner Seed wird beim Import über Electrum geprüft — Zweige mit vorhandenem Guthaben werden automatisch erkannt.",
   "psbt_join_unavailable_remote": "Zusammenführen ist im Remote-Modus nicht verfügbar",
   "feature_unavailable_remote": "Diese Funktion ist im Remote-Modus (Electrum) nicht verfügbar",
   "setup_solo_not_available_remote": "Solo-Mining benötigt einen lokalen Node — im Remote-Modus (Electrum) nicht verfügbar",
