+++ web-wallet/src/assets/locales/el.json
@@ -140,7 +140,6 @@
   "confirm_send": "Επιβεβαίωση & Αποστολή",
   "confirm_send_message": "Παρακαλώ ελέγξτε τις λεπτομέρειες της συναλλαγής πριν την αποστολή.",
   "confirm_transaction": "Επιβεβαίωση συναλλαγής",
-  "confirm_wallet_password": "Επιβεβαίωση κωδικού",
   "confirmations": "Επιβεβαιώσεις",
   "confirmations_short": "επιβ.",
   "confirmed": "Επιβεβαιωμένο",
@@ -224,7 +223,6 @@
   "edit_contact": "Επεξεργασία επαφής",
   "effective_at_block": "Σε ισχύ από το block",
   "effective_height": "Ύψος ισχύος",
-  "encrypt_wallet": "Κρυπτογράφηση πορτοφολιού με κωδικό",
   "enter_forging_address": "Εισάγετε διεύθυνση σφυρηλάτησης ή pool",
   "error": "Σφάλμα",
   "error_checking_status": "Σφάλμα ελέγχου κατάστασης",
@@ -269,10 +267,8 @@
   "import": "Εισαγωγή",
   "import_mnemonic_description": "Εισάγετε τη φράση ανάκτησης 12 ή 24 λέξεων για να επαναφέρετε το πορτοφόλι σας.",
   "import_to_wallet": "Εισαγωγή στο πορτοφόλι",
-  "import_wallet": "Εισαγωγή πορτοφολιού",
   "import_wif_description": "Εισαγωγή ιδιωτικού κλειδιού WIF ως περιγραφέα wpkh (Native SegWit) στο πορτοφόλι",
   "import_wif_wpkh": "Εισαγωγή κλειδιού WIF (wpkh)",
-  "importing": "Εισαγωγή",
   "inbound": "Εισερχόμενο",
   "incoming_payment": "Εισερχόμενη πληρωμή",
   "incorrect_word": "Λανθασμένη λέξη",
@@ -541,7 +537,6 @@
   "passphrase": "Φράση ανάκτησης",
   "passphrase_mismatch": "Οι κωδικοί δεν ταιριάζουν",
   "password": "Κωδικός",
-  "password_mismatch": "Οι κωδικοί δεν ταιριάζουν",
   "payment_confirmed": "Πληρωμή επιβεβαιώθηκε",
   "payment_uri": "URI πληρωμής",
   "payment_uri_copied": "Το URI πληρωμής αντιγράφηκε στο πρόχειρο",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "παλαιά έκδοση υπολογιστή",
   "restore_branch_pocx": "κινητό",
   "restore_history_found": "Βρέθηκαν κεφάλαια σε: {branches}",
-  "restore_no_history": "Δεν βρέθηκε προηγούμενο ιστορικό για αυτή τη φράση — το πορτοφόλι ξεκινά από την αρχή.",
   "reset": "Επαναφορά",
   "reset_mining_config": "Επαναφορά ρύθμισης εξόρυξης",
   "reset_mining_config_confirm": "Αυτό θα καθαρίσει όλες τις ρυθμίσεις εξόρυξης και θα επιστρέψει στον οδηγό ρύθμισης. Τα αρχεία plot δεν θα διαγραφούν. Είστε σίγουροι;",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Μπορείτε προαιρετικά να κρυπτογραφήσετε το πορτοφόλι σας με κωδικό. Αυτό προστατεύει το αρχείο πορτοφολιού και απαιτεί τον κωδικό για πρόσβαση στα κεφάλαιά σας.",
   "wallet_encryption_title": "Κρυπτογράφηση πορτοφολιού",
   "wallet_encryption_warning": "Αν ξεχάσετε αυτόν τον κωδικό, μπορείτε ακόμα να ανακτήσετε το πορτοφόλι σας χρησιμοποιώντας τη φράση 24 λέξεων. Ωστόσο, θα χρειαστεί να δημιουργήσετε νέο πορτοφόλι.",
-  "wallet_imported_success": "Το πορτοφόλι '{name}' εισήχθη επιτυχώς",
   "wallet_lock_status_all_locked": "Όλα τα κρυπτογραφημένα πορτοφόλια κλειδωμένα",
   "wallet_lock_status_some_unlocked": "Ξεκλειδώθηκαν {unlocked} από {total} κρυπτογραφημένα πορτοφόλια",
   "wallet_locked_message": "Αυτό το πορτοφόλι είναι κλειδωμένο. Παρακαλώ εισάγετε τον κωδικό για να το ξεκλειδώσετε προσωρινά.",
   "wallet_name": "Όνομα πορτοφολιού",
-  "wallet_name_hint": "Επιλέξτε ένα όνομα για να αναγνωρίσετε αυτό το πορτοφόλι",
   "wallet_not_encrypted": "Μη κρυπτογραφημένο",
-  "wallet_password": "Κωδικός πορτοφολιού",
   "wallet_status_notifications": "Ειδοποιήσεις κατάστασης πορτοφολιού",
   "wallets": "Πορτοφόλι(α)",
   "warn": "Προειδ.",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "συγχρονίστηκε πριν από {seconds} δ",
   "wallet_name_invalid_local": "Χρησιμοποιήστε 1–32 γράμματα, ψηφία, «-» ή «_»",
   "wallet_name_hint_local": "Μόνο γράμματα, ψηφία, «-» και «_»",
-  "wallet_encryption_info_local": "Προαιρετικά, κρυπτογραφήστε το seed του πορτοφολιού στον δίσκο με έναν κωδικό. Θα το χρειαστείτε για να ξεκλειδώσετε το πορτοφόλι για αποστολή. Η φράση ανάκτησης από μόνη της επαναφέρει πάντα τα κεφάλαιά σας.",
-  "restore_probe_info": "Το ιστορικό του seed σας ελέγχεται μέσω Electrum κατά την εισαγωγή — οι διακλαδώσεις με υπάρχοντα κεφάλαια εντοπίζονται αυτόματα.",
   "psbt_join_unavailable_remote": "Η συνένωση δεν είναι διαθέσιμη σε απομακρυσμένη λειτουργία",
   "feature_unavailable_remote": "Αυτή η λειτουργία δεν είναι διαθέσιμη σε απομακρυσμένη (Electrum) λειτουργία",
   "setup_solo_not_available_remote": "Η μεμονωμένη εξόρυξη χρειάζεται τοπικό κόμβο — δεν είναι διαθέσιμη σε απομακρυσμένη (Electrum) λειτουργία",
