+++ web-wallet/src/assets/locales/en.json
@@ -140,7 +140,6 @@
   "confirm_send": "Confirm & Send",
   "confirm_send_message": "Please review the transaction details before sending.",
   "confirm_transaction": "Confirm Transaction",
-  "confirm_wallet_password": "Confirm Password",
   "confirmations": "Confirmations",
   "confirmations_short": "conf",
   "confirmed": "Confirmed",
@@ -214,7 +213,6 @@
   "edit_contact": "Edit Contact",
   "effective_at_block": "Effective at block",
   "effective_height": "Effective Height",
-  "encrypt_wallet": "Encrypt wallet with password",
   "enter_forging_address": "Enter forging or pool address",
   "error": "Error",
   "error_checking_status": "Error checking status",
@@ -256,8 +254,6 @@
   "import_to_wallet": "Import to Wallet",
   "import_wif_description": "Import a WIF private key as a wpkh (Native SegWit) descriptor into wallet",
   "import_wif_wpkh": "Import WIF Key (wpkh)",
-  "import_wallet": "Import Wallet",
-  "importing": "Importing",
   "inbound": "Inbound",
   "incoming_payment": "Incoming Payment",
   "incorrect_word": "Incorrect word",
@@ -675,7 +671,6 @@
   "passphrase": "Recovery Phrase",
   "passphrase_mismatch": "Passphrases do not match",
   "password": "Password",
-  "password_mismatch": "Passwords do not match",
   "payment_confirmed": "Payment Confirmed",
   "payment_uri": "Payment URI",
   "payment_uri_copied": "Payment URI copied to clipboard",
@@ -709,7 +704,6 @@
   "restore_branch_legacy": "legacy desktop",
   "restore_branch_pocx": "mobile",
   "restore_history_found": "Funds found on: {branches}",
-  "restore_no_history": "No prior history was found for this phrase — the wallet starts fresh.",
   "reset": "Reset",
   "reset_to_defaults": "Reset to Defaults",
   "reset_mining_config": "Reset Mining Configuration",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "You can optionally encrypt your wallet with a password. This protects your wallet file and requires the password to access your funds.",
   "wallet_encryption_title": "Wallet Encryption",
   "wallet_encryption_warning": "If you forget this password, you can still recover your wallet using your 24-word seed phrase. However, you will need to create a new wallet.",
-  "wallet_imported_success": "Wallet '{name}' imported successfully",
   "wallet_lock_status_all_locked": "All encrypted wallets locked",
   "wallet_lock_status_some_unlocked": "{unlocked} of {total} encrypted wallets unlocked",
   "wallet_locked_message": "This wallet is locked. Please enter the passphrase to unlock it.",
   "wallet_name": "Wallet Name",
-  "wallet_name_hint": "Choose a name to identify this wallet",
   "wallet_not_encrypted": "Not encrypted",
-  "wallet_password": "Wallet Password",
   "wallet_status_notifications": "Wallet Status Notifications",
   "wallets": "Wallet(s)",
   "warn": "Warn",
@@ -1205,8 +1196,6 @@
   "electrum_synced_ago": "synced {seconds}s ago",
   "wallet_name_invalid_local": "Use 1-32 letters, digits, '-' or '_'",
   "wallet_name_hint_local": "Letters, digits, '-' and '_' only",
-  "wallet_encryption_info_local": "Optionally encrypt the wallet seed on disk with a password. You will need it to unlock the wallet for sending. The recovery phrase alone always restores your funds.",
-  "restore_probe_info": "Your seed's history is probed over Electrum during the import — branches with existing funds are detected automatically.",
   "psbt_join_unavailable_remote": "Join is not available in remote mode",
   "feature_unavailable_remote": "This feature is not available in remote (Electrum) mode",
   "setup_solo_not_available_remote": "Solo mining needs a local node — not available in remote (Electrum) mode",
