+++ web-wallet/src/assets/locales/es-es.json
@@ -140,7 +140,6 @@
   "confirm_send": "Confirmar y enviar",
   "confirm_send_message": "Por favor revise los detalles de la transacción antes de enviar.",
   "confirm_transaction": "Confirmar transacción",
-  "confirm_wallet_password": "Confirmar contraseña",
   "confirmations": "Confirmaciones",
   "confirmations_short": "conf",
   "confirmed": "Confirmado",
@@ -224,7 +223,6 @@
   "edit_contact": "Editar contacto",
   "effective_at_block": "Efectivo en el bloque",
   "effective_height": "Altura efectiva",
-  "encrypt_wallet": "Cifrar billetera con contraseña",
   "enter_forging_address": "Ingrese la dirección de forja o de pool",
   "error": "Error",
   "error_checking_status": "Error al verificar estado",
@@ -269,10 +267,8 @@
   "import": "Importar",
   "import_mnemonic_description": "Ingrese su frase de recuperación de 12 o 24 palabras para restaurar su billetera.",
   "import_to_wallet": "Importar a billetera",
-  "import_wallet": "Importar billetera",
   "import_wif_description": "Importar una clave privada WIF como descriptor wpkh (Native SegWit) a la billetera",
   "import_wif_wpkh": "Importar clave WIF (wpkh)",
-  "importing": "Importando",
   "inbound": "Entrante",
   "incoming_payment": "Pago entrante",
   "incorrect_word": "Palabra incorrecta",
@@ -541,7 +537,6 @@
   "passphrase": "Frase de recuperación",
   "passphrase_mismatch": "Las frases de contraseña no coinciden",
   "password": "Contraseña",
-  "password_mismatch": "Las contraseñas no coinciden",
   "payment_confirmed": "Pago confirmado",
   "payment_uri": "URI de pago",
   "payment_uri_copied": "URI de pago copiada al portapapeles",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "escritorio antiguo",
   "restore_branch_pocx": "móvil",
   "restore_history_found": "Fondos encontrados en: {branches}",
-  "restore_no_history": "No se encontró historial previo para esta frase — la cartera empieza de cero.",
   "reset": "Restablecer",
   "reset_mining_config": "Restablecer configuración de minería",
   "reset_mining_config_confirm": "Esto limpiará todas las configuraciones de minería y volverá al asistente de configuración. Los archivos de plot no serán eliminados. ¿Está seguro?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Opcionalmente puede cifrar su billetera con una contraseña. Esto protege el archivo de su billetera y requiere la contraseña para acceder a sus fondos.",
   "wallet_encryption_title": "Cifrado de billetera",
   "wallet_encryption_warning": "Si olvida esta contraseña, aún puede recuperar su billetera usando su frase semilla de 24 palabras. Sin embargo, necesitará crear una nueva billetera.",
-  "wallet_imported_success": "Billetera '{name}' importada exitosamente",
   "wallet_lock_status_all_locked": "Todas las billeteras cifradas bloqueadas",
   "wallet_lock_status_some_unlocked": "{unlocked} de {total} billeteras cifradas desbloqueadas",
   "wallet_locked_message": "Esta billetera está bloqueada. Por favor ingrese la frase de contraseña para desbloquearla temporalmente.",
   "wallet_name": "Nombre de billetera",
-  "wallet_name_hint": "Elija un nombre para identificar esta billetera",
   "wallet_not_encrypted": "Sin cifrar",
-  "wallet_password": "Contraseña de billetera",
   "wallet_status_notifications": "Notificaciones de estado de billetera",
   "wallets": "Billetera(s)",
   "warn": "Advert.",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "sincronizado hace {seconds}s",
   "wallet_name_invalid_local": "Usa de 1 a 32 letras, dígitos, '-' o '_'",
   "wallet_name_hint_local": "Solo letras, dígitos, '-' y '_'",
-  "wallet_encryption_info_local": "Opcionalmente, cifra la semilla de la billetera en el disco con una contraseña. La necesitarás para desbloquear la billetera al enviar. La frase de recuperación por sí sola siempre restaura tus fondos.",
-  "restore_probe_info": "El historial de tu semilla se comprueba a través de Electrum durante la importación — las ramas con fondos existentes se detectan automáticamente.",
   "psbt_join_unavailable_remote": "Combinar no está disponible en modo remoto",
   "feature_unavailable_remote": "Esta función no está disponible en modo remoto (Electrum)",
   "setup_solo_not_available_remote": "La minería en solitario necesita un nodo local — no disponible en modo remoto (Electrum)",
