+++ web-wallet/src/assets/locales/fi.json
@@ -140,7 +140,6 @@
   "confirm_send": "Vahvista ja laheta",
   "confirm_send_message": "Tarkista tapahtuman tiedot ennen lahettamista.",
   "confirm_transaction": "Vahvista tapahtuma",
-  "confirm_wallet_password": "Vahvista salasana",
   "confirmations": "Vahvistukset",
   "confirmations_short": "vahv.",
   "confirmed": "Vahvistettu",
@@ -224,7 +223,6 @@
   "edit_contact": "Muokkaa yhteystietoa",
   "effective_at_block": "Voimassa lohkosta",
   "effective_height": "Voimaantulokorkeus",
-  "encrypt_wallet": "Salaa lompakko salasanalla",
   "enter_forging_address": "Syota takontaosoite tai poolin osoite",
   "error": "Virhe",
   "error_checking_status": "Virhe tilan tarkistuksessa",
@@ -269,10 +267,8 @@
   "import": "Tuo",
   "import_mnemonic_description": "Syota 12 tai 24 sanan palautuslauseesi palauttaaksesi lompakkosi.",
   "import_to_wallet": "Tuo lompakkoon",
-  "import_wallet": "Tuo lompakko",
   "import_wif_description": "Tuo WIF-yksityisavain wpkh (Native SegWit) -kuvaajana lompakkoon",
   "import_wif_wpkh": "Tuo WIF-avain (wpkh)",
-  "importing": "Tuodaan",
   "inbound": "Saapuva",
   "incoming_payment": "Saapuva maksu",
   "incorrect_word": "Vaara sana",
@@ -541,7 +537,6 @@
   "passphrase": "Palautuslause",
   "passphrase_mismatch": "Salasanat eivat tasmaa",
   "password": "Salasana",
-  "password_mismatch": "Salasanat eivat tasmaa",
   "payment_confirmed": "Maksu vahvistettu",
   "payment_uri": "Maksu-URI",
   "payment_uri_copied": "Maksu-URI kopioitu leikepoydale",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "vanha työpöytä",
   "restore_branch_pocx": "mobiili",
   "restore_history_found": "Varoja löytyi: {branches}",
-  "restore_no_history": "Tälle lauseelle ei löytynyt aiempaa historiaa — lompakko aloittaa alusta.",
   "reset": "Nollaa",
   "reset_mining_config": "Nollaa louhinta-asetukset",
   "reset_mining_config_confirm": "Tama tyhjentaa kaikki louhinta-asetukset ja palaa ohjattuun asetukseen. Plot-tiedostoja ei poisteta. Oletko varma?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Voit valinnaisesti salata lompakkosi salasanalla. Tama suojaa lompakkotiedostoasi ja vaatii salasanan varojen kayttamiseen.",
   "wallet_encryption_title": "Lompakon salaus",
   "wallet_encryption_warning": "Jos unohdat taman salasanan, voit silti palauttaa lompakkosi kayttamalla 24 sanan siemenlausettasi. Sinun on kuitenkin luotava uusi lompakko.",
-  "wallet_imported_success": "Lompakko '{name}' tuotu onnistuneesti",
   "wallet_lock_status_all_locked": "Kaikki salatut lompakot lukittu",
   "wallet_lock_status_some_unlocked": "{unlocked} / {total} salatuista lompakoista avattu",
   "wallet_locked_message": "Tama lompakko on lukittu. Syota salasana avataksesi sen valikaikaisesti.",
   "wallet_name": "Lompakon nimi",
-  "wallet_name_hint": "Valitse nimi tunnistaaksesi taman lompakon",
   "wallet_not_encrypted": "Ei salattu",
-  "wallet_password": "Lompakon salasana",
   "wallet_status_notifications": "Lompakon tilailmoitukset",
   "wallets": "Lompakko(t)",
   "warn": "Varoitus",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "synkronoitu {seconds} s sitten",
   "wallet_name_invalid_local": "Käytä 1–32 kirjainta, numeroa, ”-” tai ”_”",
   "wallet_name_hint_local": "Vain kirjaimia, numeroita, ”-” ja ”_”",
-  "wallet_encryption_info_local": "Voit halutessasi salata lompakon siemenen levyllä salasanalla. Tarvitset sitä lompakon lukituksen avaamiseen lähettämistä varten. Pelkkä palautuslause palauttaa aina varasi.",
-  "restore_probe_info": "Siemenesi historia tutkitaan Electrumin kautta tuonnin aikana — haarat, joissa on varoja, tunnistetaan automaattisesti.",
   "psbt_join_unavailable_remote": "Yhdistäminen ei ole käytettävissä etätilassa",
   "feature_unavailable_remote": "Tämä ominaisuus ei ole käytettävissä etätilassa (Electrum)",
   "setup_solo_not_available_remote": "Yksinlouhinta tarvitsee paikallisen solmun — ei käytettävissä etätilassa (Electrum)",
