+++ web-wallet/src/assets/locales/fr.json
@@ -140,7 +140,6 @@
   "confirm_send": "Confirmer et envoyer",
   "confirm_send_message": "Veuillez verifier les details de la transaction avant d'envoyer.",
   "confirm_transaction": "Confirmer la transaction",
-  "confirm_wallet_password": "Confirmer le mot de passe",
   "confirmations": "Confirmations",
   "confirmations_short": "conf",
   "confirmed": "Confirme",
@@ -224,7 +223,6 @@
   "edit_contact": "Modifier le contact",
   "effective_at_block": "Effectif au bloc",
   "effective_height": "Hauteur d'effet",
-  "encrypt_wallet": "Chiffrer le portefeuille avec un mot de passe",
   "enter_forging_address": "Entrer l'adresse de forging ou de pool",
   "error": "Erreur",
   "error_checking_status": "Erreur lors de la verification du statut",
@@ -269,10 +267,8 @@
   "import": "Importer",
   "import_mnemonic_description": "Entrez votre phrase de recuperation de 12 ou 24 mots pour restaurer votre portefeuille.",
   "import_to_wallet": "Importer dans le portefeuille",
-  "import_wallet": "Importer un portefeuille",
   "import_wif_description": "Importer une cle privee WIF en tant que descripteur wpkh (Native SegWit) dans le portefeuille",
   "import_wif_wpkh": "Importer une cle WIF (wpkh)",
-  "importing": "Importation",
   "inbound": "Entrant",
   "incoming_payment": "Paiement entrant",
   "incorrect_word": "Mot incorrect",
@@ -541,7 +537,6 @@
   "passphrase": "Phrase de recuperation",
   "passphrase_mismatch": "Les phrases de passe ne correspondent pas",
   "password": "Mot de passe",
-  "password_mismatch": "Les mots de passe ne correspondent pas",
   "payment_confirmed": "Paiement confirme",
   "payment_uri": "URI de paiement",
   "payment_uri_copied": "URI de paiement copiee dans le presse-papiers",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "bureau hérité",
   "restore_branch_pocx": "mobile",
   "restore_history_found": "Fonds trouvés sur : {branches}",
-  "restore_no_history": "Aucun historique antérieur trouvé pour cette phrase — le portefeuille démarre à neuf.",
   "reset": "Reinitialiser",
   "reset_mining_config": "Reinitialiser la configuration de minage",
   "reset_mining_config_confirm": "Ceci effacera tous les parametres de minage et retournera a l'assistant de configuration. Les fichiers plot ne seront pas supprimes. Etes-vous sur ?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Vous pouvez optionnellement chiffrer votre portefeuille avec un mot de passe. Cela protege votre fichier de portefeuille et necessite le mot de passe pour acceder a vos fonds.",
   "wallet_encryption_title": "Chiffrement du portefeuille",
   "wallet_encryption_warning": "Si vous oubliez ce mot de passe, vous pouvez toujours recuperer votre portefeuille en utilisant votre phrase de 24 mots. Cependant, vous devrez creer un nouveau portefeuille.",
-  "wallet_imported_success": "Portefeuille '{name}' importe avec succes",
   "wallet_lock_status_all_locked": "Tous les portefeuilles chiffres verrouilles",
   "wallet_lock_status_some_unlocked": "{unlocked} sur {total} portefeuilles chiffres deverrouilles",
   "wallet_locked_message": "Ce portefeuille est verrouille. Veuillez entrer le mot de passe pour le deverrouiller temporairement.",
   "wallet_name": "Nom du portefeuille",
-  "wallet_name_hint": "Choisissez un nom pour identifier ce portefeuille",
   "wallet_not_encrypted": "Non chiffre",
-  "wallet_password": "Mot de passe du portefeuille",
   "wallet_status_notifications": "Notifications de statut du portefeuille",
   "wallets": "Portefeuille(s)",
   "warn": "Avert.",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "synchronisé il y a {seconds}s",
   "wallet_name_invalid_local": "Utilisez de 1 à 32 lettres, chiffres, '-' ou '_'",
   "wallet_name_hint_local": "Lettres, chiffres, '-' et '_' uniquement",
-  "wallet_encryption_info_local": "Chiffrez éventuellement la graine du portefeuille sur le disque avec un mot de passe. Il sera nécessaire pour déverrouiller le portefeuille lors de l'envoi. La phrase de récupération seule restaure toujours vos fonds.",
-  "restore_probe_info": "L'historique de votre graine est sondé via Electrum lors de l'import — les branches contenant des fonds existants sont détectées automatiquement.",
   "psbt_join_unavailable_remote": "La fusion n'est pas disponible en mode distant",
   "feature_unavailable_remote": "Cette fonctionnalité n'est pas disponible en mode distant (Electrum)",
   "setup_solo_not_available_remote": "Le minage en solo nécessite un nœud local — non disponible en mode distant (Electrum)",
