+++ web-wallet/src/assets/locales/gl.json
@@ -140,7 +140,6 @@
   "confirm_send": "Confirmar e enviar",
   "confirm_send_message": "Por favor, revisa os detalles da transacción antes de enviar.",
   "confirm_transaction": "Confirmar transacción",
-  "confirm_wallet_password": "Confirmar contrasinal",
   "confirmations": "Confirmacións",
   "confirmations_short": "conf",
   "confirmed": "Confirmado",
@@ -224,7 +223,6 @@
   "edit_contact": "Editar contacto",
   "effective_at_block": "Efectivo no bloque",
   "effective_height": "Altura efectiva",
-  "encrypt_wallet": "Cifrar carteira con contrasinal",
   "enter_forging_address": "Introduce o enderezo de forxa ou de pool",
   "error": "Erro",
   "error_checking_status": "Erro ao verificar o estado",
@@ -269,10 +267,8 @@
   "import": "Importar",
   "import_mnemonic_description": "Introduce a túa frase de recuperación de 12 ou 24 palabras para restaurar a túa carteira.",
   "import_to_wallet": "Importar á carteira",
-  "import_wallet": "Importar carteira",
   "import_wif_description": "Importar unha chave privada WIF como un descritor wpkh (Native SegWit) na carteira",
   "import_wif_wpkh": "Importar chave WIF (wpkh)",
-  "importing": "Importando",
   "inbound": "Entrante",
   "incoming_payment": "Pago entrante",
   "incorrect_word": "Palabra incorrecta",
@@ -541,7 +537,6 @@
   "passphrase": "Frase de recuperación",
   "passphrase_mismatch": "As frases de contrasinal non coinciden",
   "password": "Contrasinal",
-  "password_mismatch": "Os contrasinais non coinciden",
   "payment_confirmed": "Pago confirmado",
   "payment_uri": "URI de pago",
   "payment_uri_copied": "URI de pago copiada ao portapapeis",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "escritorio antigo",
   "restore_branch_pocx": "móbil",
   "restore_history_found": "Atopáronse fondos en: {branches}",
-  "restore_no_history": "Non se atopou historial previo para esta frase — a carteira comeza de cero.",
   "reset": "Restablecer",
   "reset_mining_config": "Restablecer configuración de minaría",
   "reset_mining_config_confirm": "Isto borrará todas as configuracións de minaría e volverá ao asistente de configuración. Os ficheiros de trazado non se eliminarán. Estás seguro?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Opcionalmente podes cifrar a túa carteira cunha contrasinal. Isto protexe o ficheiro da túa carteira e require a contrasinal para acceder aos teus fondos.",
   "wallet_encryption_title": "Cifrado de carteira",
   "wallet_encryption_warning": "Se esqueces esta contrasinal, aínda poderás recuperar a túa carteira usando a túa frase semente de 24 palabras. Non obstante, terás que crear unha nova carteira.",
-  "wallet_imported_success": "Carteira '{name}' importada con éxito",
   "wallet_lock_status_all_locked": "Todas as carteiras cifradas bloqueadas",
   "wallet_lock_status_some_unlocked": "{unlocked} de {total} carteiras cifradas desbloqueadas",
   "wallet_locked_message": "Esta carteira está bloqueada. Por favor, introduce a frase de contrasinal para desbloqueala temporalmente.",
   "wallet_name": "Nome da carteira",
-  "wallet_name_hint": "Escolle un nome para identificar esta carteira",
   "wallet_not_encrypted": "Sen cifrar",
-  "wallet_password": "Contrasinal da carteira",
   "wallet_status_notifications": "Notificacións de estado da carteira",
   "wallets": "Carteira(s)",
   "warn": "Aviso",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "sincronizado hai {seconds}s",
   "wallet_name_invalid_local": "Usa de 1 a 32 letras, díxitos, '-' ou '_'",
   "wallet_name_hint_local": "Só letras, díxitos, '-' e '_'",
-  "wallet_encryption_info_local": "Opcionalmente, cifra a semente da carteira no disco cun contrasinal. Fará falta para desbloquear a carteira ao enviar. A frase de recuperación por si soa sempre restaura os teus fondos.",
-  "restore_probe_info": "O historial da túa semente sóndase mediante Electrum durante a importación — as ramas con fondos existentes detéctanse automaticamente.",
   "psbt_join_unavailable_remote": "Combinar non está dispoñible no modo remoto",
   "feature_unavailable_remote": "Esta función non está dispoñible no modo remoto (Electrum)",
   "setup_solo_not_available_remote": "A minaría en solitario precisa dun nodo local — non dispoñible no modo remoto (Electrum)",
