+++ web-wallet/src/assets/locales/hi.json
@@ -140,7 +140,6 @@
   "confirm_send": "पुष्टि करें और भेजें",
   "confirm_send_message": "भेजने से पहले लेनदेन विवरण की समीक्षा करें।",
   "confirm_transaction": "लेनदेन की पुष्टि करें",
-  "confirm_wallet_password": "पासवर्ड की पुष्टि करें",
   "confirmations": "पुष्टिकरण",
   "confirmations_short": "पुष्टि",
   "confirmed": "पुष्टि हुई",
@@ -224,7 +223,6 @@
   "edit_contact": "संपर्क संपादित करें",
   "effective_at_block": "ब्लॉक पर प्रभावी",
   "effective_height": "प्रभावी ऊंचाई",
-  "encrypt_wallet": "पासवर्ड से वॉलेट एन्क्रिप्ट करें",
   "enter_forging_address": "फोर्जिंग या पूल पता दर्ज करें",
   "error": "त्रुटि",
   "error_checking_status": "स्थिति जांचने में त्रुटि",
@@ -269,10 +267,8 @@
   "import": "आयात करें",
   "import_mnemonic_description": "अपना वॉलेट पुनर्स्थापित करने के लिए अपना 12 या 24 शब्द रिकवरी वाक्यांश दर्ज करें।",
   "import_to_wallet": "वॉलेट में आयात करें",
-  "import_wallet": "वॉलेट आयात करें",
   "import_wif_description": "WIF प्राइवेट की को wpkh (Native SegWit) डिस्क्रिप्टर के रूप में वॉलेट में आयात करें",
   "import_wif_wpkh": "WIF की आयात करें (wpkh)",
-  "importing": "आयात हो रहा है",
   "inbound": "इनबाउंड",
   "incoming_payment": "आने वाला भुगतान",
   "incorrect_word": "गलत शब्द",
@@ -541,7 +537,6 @@
   "passphrase": "रिकवरी वाक्यांश",
   "passphrase_mismatch": "पासफ्रेज मेल नहीं खाते",
   "password": "पासवर्ड",
-  "password_mismatch": "पासवर्ड मेल नहीं खाते",
   "payment_confirmed": "भुगतान पुष्टि हुआ",
   "payment_uri": "भुगतान URI",
   "payment_uri_copied": "भुगतान URI क्लिपबोर्ड पर कॉपी किया गया",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "पुराना डेस्कटॉप",
   "restore_branch_pocx": "मोबाइल",
   "restore_history_found": "धनराशि यहाँ मिली: {branches}",
-  "restore_no_history": "इस वाक्यांश के लिए कोई पुराना इतिहास नहीं मिला — वॉलेट नए सिरे से शुरू होगा।",
   "reset": "रीसेट",
   "reset_mining_config": "माइनिंग कॉन्फ़िगरेशन रीसेट करें",
   "reset_mining_config_confirm": "यह सभी माइनिंग सेटिंग्स साफ कर देगा और सेटअप विज़ार्ड पर वापस ले जाएगा। प्लॉट फ़ाइलें हटाई नहीं जाएंगी। क्या आप सुनिश्चित हैं?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "आप वैकल्पिक रूप से अपने वॉलेट को पासवर्ड से एन्क्रिप्ट कर सकते हैं। यह आपकी वॉलेट फाइल की सुरक्षा करता है और आपके फंड तक पहुंचने के लिए पासवर्ड की आवश्यकता होती है।",
   "wallet_encryption_title": "वॉलेट एन्क्रिप्शन",
   "wallet_encryption_warning": "यदि आप यह पासवर्ड भूल जाते हैं, तो भी आप अपने 24-शब्द सीड वाक्यांश का उपयोग करके अपना वॉलेट पुनर्प्राप्त कर सकते हैं। हालांकि, आपको एक नया वॉलेट बनाना होगा।",
-  "wallet_imported_success": "वॉलेट '{name}' सफलतापूर्वक आयात किया गया",
   "wallet_lock_status_all_locked": "सभी एन्क्रिप्टेड वॉलेट लॉक",
   "wallet_lock_status_some_unlocked": "{total} में से {unlocked} एन्क्रिप्टेड वॉलेट अनलॉक",
   "wallet_locked_message": "यह वॉलेट लॉक है। अस्थायी रूप से अनलॉक करने के लिए पासफ्रेज दर्ज करें।",
   "wallet_name": "वॉलेट नाम",
-  "wallet_name_hint": "इस वॉलेट की पहचान के लिए एक नाम चुनें",
   "wallet_not_encrypted": "एन्क्रिप्टेड नहीं",
-  "wallet_password": "वॉलेट पासवर्ड",
   "wallet_status_notifications": "वॉलेट स्थिति सूचनाएं",
   "wallets": "वॉलेट",
   "warn": "चेता.",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "{seconds} सेकंड पहले सिंक किया गया",
   "wallet_name_invalid_local": "1–32 अक्षर, अंक, '-' या '_' का उपयोग करें",
   "wallet_name_hint_local": "केवल अक्षर, अंक, '-' और '_'",
-  "wallet_encryption_info_local": "वैकल्पिक रूप से वॉलेट सीड को डिस्क पर पासवर्ड से एन्क्रिप्ट करें। भेजने के लिए वॉलेट अनलॉक करने हेतु आपको इसकी आवश्यकता होगी। केवल रिकवरी वाक्यांश हमेशा आपके धन को पुनर्स्थापित करता है।",
-  "restore_probe_info": "आयात के दौरान आपकी सीड का इतिहास Electrum पर जांचा जाता है — मौजूदा धन वाली शाखाएं स्वचालित रूप से पहचानी जाती हैं।",
   "psbt_join_unavailable_remote": "रिमोट मोड में जॉइन उपलब्ध नहीं है",
   "feature_unavailable_remote": "यह सुविधा रिमोट (Electrum) मोड में उपलब्ध नहीं है",
   "setup_solo_not_available_remote": "सोलो माइनिंग के लिए स्थानीय नोड आवश्यक है — रिमोट (Electrum) मोड में उपलब्ध नहीं है",
