+++ web-wallet/src/assets/locales/hr.json
@@ -140,7 +140,6 @@
   "confirm_send": "Potvrdi i posalji",
   "confirm_send_message": "Molimo provjerite detalje transakcije prije slanja.",
   "confirm_transaction": "Potvrdi transakciju",
-  "confirm_wallet_password": "Potvrdi lozinku",
   "confirmations": "Potvrde",
   "confirmations_short": "potv.",
   "confirmed": "Potvrdeno",
@@ -224,7 +223,6 @@
   "edit_contact": "Uredi kontakt",
   "effective_at_block": "Na snazi od bloka",
   "effective_height": "Efektivna visina",
-  "encrypt_wallet": "Sifriraj novcanik lozinkom",
   "enter_forging_address": "Unesite adresu kovanja ili poola",
   "error": "Greska",
   "error_checking_status": "Greska pri provjeri statusa",
@@ -269,10 +267,8 @@
   "import": "Uvezi",
   "import_mnemonic_description": "Unesite frazu za oporavak od 12 ili 24 rijeci kako biste vratili novcanik.",
   "import_to_wallet": "Uvezi u novcanik",
-  "import_wallet": "Uvezi novcanik",
   "import_wif_description": "Uvezite WIF privatni kljuc kao wpkh (Native SegWit) deskriptor u novcanik",
   "import_wif_wpkh": "Uvezi WIF kljuc (wpkh)",
-  "importing": "Uvoz",
   "inbound": "Dolazno",
   "incoming_payment": "Dolazna uplata",
   "incorrect_word": "Netocna rijec",
@@ -541,7 +537,6 @@
   "passphrase": "Fraza za oporavak",
   "passphrase_mismatch": "Lozinke se ne podudaraju",
   "password": "Lozinka",
-  "password_mismatch": "Lozinke se ne podudaraju",
   "payment_confirmed": "Uplata potvrdena",
   "payment_uri": "URI za placanje",
   "payment_uri_copied": "URI za placanje kopiran u meduspremnik",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "stari desktop",
   "restore_branch_pocx": "mobilni",
   "restore_history_found": "Sredstva pronađena na: {branches}",
-  "restore_no_history": "Za ovu frazu nije pronađena ranija povijest — novčanik počinje ispočetka.",
   "reset": "Resetiraj",
   "reset_mining_config": "Resetiraj konfiguraciju rudarenja",
   "reset_mining_config_confirm": "Ovo ce obrisati sve postavke rudarenja i vratiti se na carobnjak za postavljanje. Plot datoteke nece biti obrisane. Jeste li sigurni?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Mozete opcionalno sifrirati novcanik lozinkom. Ovo stiti datoteku novcanika i zahtijeva lozinku za pristup sredstvima.",
   "wallet_encryption_title": "Sifriranje novcanika",
   "wallet_encryption_warning": "Ako zaboravite ovu lozinku, jos uvijek mozete oporaviti novcanik koristeci frazu od 24 rijeci. Medutim, morat cete kreirati novi novcanik.",
-  "wallet_imported_success": "Novcanik '{name}' uspjesno uvezen",
   "wallet_lock_status_all_locked": "Svi sifrirani novcanici zakljucani",
   "wallet_lock_status_some_unlocked": "Otkljucano {unlocked} od {total} sifriranih novcanika",
   "wallet_locked_message": "Ovaj novcanik je zakljucan. Molimo unesite lozinku za privremeno otkljucavanje.",
   "wallet_name": "Ime novcanika",
-  "wallet_name_hint": "Odaberite ime za identifikaciju ovog novcanika",
   "wallet_not_encrypted": "Nesifrirano",
-  "wallet_password": "Lozinka novcanika",
   "wallet_status_notifications": "Obavijesti o statusu novcanika",
   "wallets": "Novcanik(ci)",
   "warn": "Upoz.",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "sinkronizirano prije {seconds} s",
   "wallet_name_invalid_local": "Koristite 1–32 slova, znamenke, „-” ili „_”",
   "wallet_name_hint_local": "Samo slova, znamenke, „-” i „_”",
-  "wallet_encryption_info_local": "Po želji šifrirajte sjeme novčanika na disku lozinkom. Trebat će vam za otključavanje novčanika radi slanja. Sama fraza za oporavak uvijek vraća vaša sredstva.",
-  "restore_probe_info": "Povijest vašeg sjemena provjerava se putem Electruma tijekom uvoza — grane s postojećim sredstvima automatski se otkrivaju.",
   "psbt_join_unavailable_remote": "Spajanje nije dostupno u udaljenom načinu rada",
   "feature_unavailable_remote": "Ova značajka nije dostupna u udaljenom (Electrum) načinu rada",
   "setup_solo_not_available_remote": "Solo rudarenje treba lokalni čvor — nije dostupno u udaljenom (Electrum) načinu rada",
