+++ web-wallet/src/assets/locales/id.json
@@ -140,7 +140,6 @@
   "confirm_send": "Konfirmasi & Kirim",
   "confirm_send_message": "Tinjau detail transaksi sebelum mengirim.",
   "confirm_transaction": "Konfirmasi Transaksi",
-  "confirm_wallet_password": "Konfirmasi Kata Sandi",
   "confirmations": "Konfirmasi",
   "confirmations_short": "konf",
   "confirmed": "Dikonfirmasi",
@@ -224,7 +223,6 @@
   "edit_contact": "Edit Kontak",
   "effective_at_block": "Berlaku pada blok",
   "effective_height": "Tinggi efektif",
-  "encrypt_wallet": "Enkripsi dompet dengan kata sandi",
   "enter_forging_address": "Masukkan alamat forging atau pool",
   "error": "Kesalahan",
   "error_checking_status": "Error memeriksa status",
@@ -269,10 +267,8 @@
   "import": "Impor",
   "import_mnemonic_description": "Masukkan frasa pemulihan 12 atau 24 kata untuk memulihkan dompet Anda.",
   "import_to_wallet": "Impor ke Dompet",
-  "import_wallet": "Impor Dompet",
   "import_wif_description": "Impor kunci privat WIF sebagai deskriptor wpkh (Native SegWit) ke dompet",
   "import_wif_wpkh": "Impor Kunci WIF (wpkh)",
-  "importing": "Mengimpor",
   "inbound": "Masuk",
   "incoming_payment": "Pembayaran Masuk",
   "incorrect_word": "Kata salah",
@@ -541,7 +537,6 @@
   "passphrase": "Frasa Pemulihan",
   "passphrase_mismatch": "Frasa sandi tidak cocok",
   "password": "Kata Sandi",
-  "password_mismatch": "Kata sandi tidak cocok",
   "payment_confirmed": "Pembayaran Dikonfirmasi",
   "payment_uri": "URI Pembayaran",
   "payment_uri_copied": "URI Pembayaran disalin ke clipboard",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "desktop lama",
   "restore_branch_pocx": "seluler",
   "restore_history_found": "Dana ditemukan di: {branches}",
-  "restore_no_history": "Tidak ada riwayat sebelumnya untuk frasa ini — dompet mulai dari awal.",
   "reset": "Atur Ulang",
   "reset_mining_config": "Reset Konfigurasi Penambangan",
   "reset_mining_config_confirm": "Ini akan menghapus semua pengaturan penambangan dan kembali ke wizard pengaturan. File plot tidak akan dihapus. Apakah Anda yakin?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Anda dapat secara opsional mengenkripsi dompet Anda dengan kata sandi. Ini melindungi file dompet Anda dan memerlukan kata sandi untuk mengakses dana.",
   "wallet_encryption_title": "Enkripsi Dompet",
   "wallet_encryption_warning": "Jika Anda lupa kata sandi ini, Anda masih dapat memulihkan dompet Anda menggunakan frasa seed 24 kata Anda. Namun, Anda perlu membuat dompet baru.",
-  "wallet_imported_success": "Dompet '{name}' berhasil diimpor",
   "wallet_lock_status_all_locked": "Semua dompet terenkripsi terkunci",
   "wallet_lock_status_some_unlocked": "{unlocked} dari {total} dompet terenkripsi tidak terkunci",
   "wallet_locked_message": "Dompet ini terkunci. Masukkan frasa sandi untuk membuka kunci sementara.",
   "wallet_name": "Nama Dompet",
-  "wallet_name_hint": "Pilih nama untuk mengidentifikasi dompet ini",
   "wallet_not_encrypted": "Tidak Terenkripsi",
-  "wallet_password": "Kata Sandi Dompet",
   "wallet_status_notifications": "Notifikasi Status Dompet",
   "wallets": "Dompet",
   "warn": "Peringatan",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "disinkronkan {seconds} dtk lalu",
   "wallet_name_invalid_local": "Gunakan 1–32 huruf, angka, '-' atau '_'",
   "wallet_name_hint_local": "Hanya huruf, angka, '-' dan '_'",
-  "wallet_encryption_info_local": "Secara opsional enkripsi seed dompet di disk dengan kata sandi. Anda akan membutuhkannya untuk membuka kunci dompet saat mengirim. Frasa pemulihan saja selalu memulihkan dana Anda.",
-  "restore_probe_info": "Riwayat seed Anda diperiksa melalui Electrum selama impor — cabang dengan dana yang ada terdeteksi secara otomatis.",
   "psbt_join_unavailable_remote": "Gabung tidak tersedia dalam mode jarak jauh",
   "feature_unavailable_remote": "Fitur ini tidak tersedia dalam mode jarak jauh (Electrum)",
   "setup_solo_not_available_remote": "Penambangan solo memerlukan node lokal — tidak tersedia dalam mode jarak jauh (Electrum)",
