+++ web-wallet/src/assets/locales/it.json
@@ -140,7 +140,6 @@
   "confirm_send": "Conferma e invia",
   "confirm_send_message": "Verifica i dettagli della transazione prima di inviare.",
   "confirm_transaction": "Conferma transazione",
-  "confirm_wallet_password": "Conferma password",
   "confirmations": "Conferme",
   "confirmations_short": "conf",
   "confirmed": "Confermato",
@@ -224,7 +223,6 @@
   "edit_contact": "Modifica contatto",
   "effective_at_block": "Effettivo al blocco",
   "effective_height": "Altezza di efficacia",
-  "encrypt_wallet": "Cripta portafoglio con password",
   "enter_forging_address": "Inserisci indirizzo forging o pool",
   "error": "Errore",
   "error_checking_status": "Errore nella verifica dello stato",
@@ -269,10 +267,8 @@
   "import": "Importa",
   "import_mnemonic_description": "Inserisci la tua frase di recupero di 12 o 24 parole per ripristinare il tuo portafoglio.",
   "import_to_wallet": "Importa nel portafoglio",
-  "import_wallet": "Importa portafoglio",
   "import_wif_description": "Importa una chiave privata WIF come descrittore wpkh (Native SegWit) nel portafoglio",
   "import_wif_wpkh": "Importa chiave WIF (wpkh)",
-  "importing": "Importazione",
   "inbound": "In entrata",
   "incoming_payment": "Pagamento in entrata",
   "incorrect_word": "Parola errata",
@@ -541,7 +537,6 @@
   "passphrase": "Frase di recupero",
   "passphrase_mismatch": "Le passphrase non corrispondono",
   "password": "Password",
-  "password_mismatch": "Le password non corrispondono",
   "payment_confirmed": "Pagamento confermato",
   "payment_uri": "URI di pagamento",
   "payment_uri_copied": "URI di pagamento copiata negli appunti",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "desktop legacy",
   "restore_branch_pocx": "mobile",
   "restore_history_found": "Fondi trovati su: {branches}",
-  "restore_no_history": "Nessuna cronologia precedente trovata per questa frase — il portafoglio parte da zero.",
   "reset": "Ripristina",
   "reset_mining_config": "Ripristina configurazione mining",
   "reset_mining_config_confirm": "Questo cancellera tutte le impostazioni di mining e tornera alla procedura guidata. I file plot non verranno eliminati. Sei sicuro?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Puoi opzionalmente criptare il tuo portafoglio con una password. Questo protegge il file del tuo portafoglio e richiede la password per accedere ai tuoi fondi.",
   "wallet_encryption_title": "Crittografia portafoglio",
   "wallet_encryption_warning": "Se dimentichi questa password, puoi comunque recuperare il tuo portafoglio usando la tua frase di 24 parole. Tuttavia, dovrai creare un nuovo portafoglio.",
-  "wallet_imported_success": "Portafoglio '{name}' importato con successo",
   "wallet_lock_status_all_locked": "Tutti i portafogli cifrati bloccati",
   "wallet_lock_status_some_unlocked": "{unlocked} di {total} portafogli cifrati sbloccati",
   "wallet_locked_message": "Questo portafoglio e bloccato. Inserisci la password per sbloccarlo temporaneamente.",
   "wallet_name": "Nome portafoglio",
-  "wallet_name_hint": "Scegli un nome per identificare questo portafoglio",
   "wallet_not_encrypted": "Non criptato",
-  "wallet_password": "Password portafoglio",
   "wallet_status_notifications": "Notifiche stato portafoglio",
   "wallets": "Portafoglio/i",
   "warn": "Avviso",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "sincronizzato {seconds}s fa",
   "wallet_name_invalid_local": "Usa da 1 a 32 lettere, cifre, '-' o '_'",
   "wallet_name_hint_local": "Solo lettere, cifre, '-' e '_'",
-  "wallet_encryption_info_local": "Facoltativamente cifra il seed del portafoglio sul disco con una password. Ti servirà per sbloccare il portafoglio per l'invio. La frase di recupero da sola ripristina sempre i tuoi fondi.",
-  "restore_probe_info": "La cronologia del tuo seed viene sondata tramite Electrum durante l'importazione — i rami con fondi esistenti vengono rilevati automaticamente.",
   "psbt_join_unavailable_remote": "L'unione non è disponibile in modalità remota",
   "feature_unavailable_remote": "Questa funzione non è disponibile in modalità remota (Electrum)",
   "setup_solo_not_available_remote": "Il mining in solo richiede un nodo locale — non disponibile in modalità remota (Electrum)",
