+++ web-wallet/src/assets/locales/ja.json
@@ -140,7 +140,6 @@
   "confirm_send": "確認して送信",
   "confirm_send_message": "送信前にトランザクションの詳細を確認してください。",
   "confirm_transaction": "トランザクションを確認",
-  "confirm_wallet_password": "パスワードを確認",
   "confirmations": "確認",
   "confirmations_short": "確認",
   "confirmed": "確認済み",
@@ -224,7 +223,6 @@
   "edit_contact": "連絡先を編集",
   "effective_at_block": "有効になるブロック",
   "effective_height": "発効ブロック高",
-  "encrypt_wallet": "パスワードでウォレットを暗号化",
   "enter_forging_address": "フォージングまたはプールアドレスを入力",
   "error": "エラー",
   "error_checking_status": "ステータス確認中にエラーが発生しました",
@@ -269,10 +267,8 @@
   "import": "インポート",
   "import_mnemonic_description": "ウォレットを復元するには、12語または24語のリカバリーフレーズを入力してください。",
   "import_to_wallet": "ウォレットにインポート",
-  "import_wallet": "ウォレットをインポート",
   "import_wif_description": "WIF秘密鍵をwpkh（ネイティブSegWit）ディスクリプタとしてウォレットにインポートします",
   "import_wif_wpkh": "WIFキーをインポート（wpkh）",
-  "importing": "インポート中",
   "inbound": "インバウンド",
   "incoming_payment": "入金",
   "incorrect_word": "単語が正しくありません",
@@ -541,7 +537,6 @@
   "passphrase": "リカバリーフレーズ",
   "passphrase_mismatch": "パスフレーズが一致しません",
   "password": "パスワード",
-  "password_mismatch": "パスワードが一致しません",
   "payment_confirmed": "支払い確認済み",
   "payment_uri": "支払いURI",
   "payment_uri_copied": "支払いURIをクリップボードにコピーしました",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "旧デスクトップ",
   "restore_branch_pocx": "モバイル",
   "restore_history_found": "資金が見つかりました: {branches}",
-  "restore_no_history": "このフレーズの過去の履歴は見つかりませんでした — ウォレットは新規に開始します。",
   "reset": "リセット",
   "reset_mining_config": "マイニング設定をリセット",
   "reset_mining_config_confirm": "これによりすべてのマイニング設定がクリアされ、セットアップウィザードに戻ります。プロットファイルは削除されません。よろしいですか？",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "オプションでパスワードを使用してウォレットを暗号化できます。これによりウォレットファイルが保護され、資金にアクセスするにはパスワードが必要になります。",
   "wallet_encryption_title": "ウォレット暗号化",
   "wallet_encryption_warning": "このパスワードを忘れた場合でも、24語のシードフレーズを使用してウォレットを復元できます。ただし、新しいウォレットを作成する必要があります。",
-  "wallet_imported_success": "ウォレット '{name}' が正常にインポートされました",
   "wallet_lock_status_all_locked": "すべての暗号化されたウォレットがロックされています",
   "wallet_lock_status_some_unlocked": "{total} 個の暗号化されたウォレットのうち {unlocked} 個がロック解除",
   "wallet_locked_message": "このウォレットはロックされています。一時的にロック解除するにはパスフレーズを入力してください。",
   "wallet_name": "ウォレット名",
-  "wallet_name_hint": "このウォレットを識別する名前を選択してください",
   "wallet_not_encrypted": "暗号化されていません",
-  "wallet_password": "ウォレットパスワード",
   "wallet_status_notifications": "ウォレットステータス通知",
   "wallets": "ウォレット",
   "warn": "警告",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "{seconds} 秒前に同期",
   "wallet_name_invalid_local": "1〜32 文字の英字、数字、「-」または「_」を使用してください",
   "wallet_name_hint_local": "英字、数字、「-」および「_」のみ",
-  "wallet_encryption_info_local": "任意で、ディスク上のウォレットシードをパスワードで暗号化できます。送金時にウォレットのロック解除に必要になります。リカバリーフレーズだけでも常に資金を復元できます。",
-  "restore_probe_info": "インポート中に、シードの履歴が Electrum を通じて調べられます — 資金のある分岐は自動的に検出されます。",
   "psbt_join_unavailable_remote": "結合はリモートモードでは利用できません",
   "feature_unavailable_remote": "この機能はリモート（Electrum）モードでは利用できません",
   "setup_solo_not_available_remote": "ソロマイニングにはローカルノードが必要です — リモート（Electrum）モードでは利用できません",
