+++ web-wallet/src/assets/locales/lt.json
@@ -140,7 +140,6 @@
   "confirm_send": "Patvirtinti ir siusti",
   "confirm_send_message": "Prasome patikrinti operacijos detales pries siuntima.",
   "confirm_transaction": "Patvirtinti operacija",
-  "confirm_wallet_password": "Patvirtinti slaptazodi",
   "confirmations": "Patvirtinimai",
   "confirmations_short": "patv.",
   "confirmed": "Patvirtinta",
@@ -224,7 +223,6 @@
   "edit_contact": "Redaguoti kontakta",
   "effective_at_block": "Isigalioja bloke",
   "effective_height": "Įsigaliojimo aukštis",
-  "encrypt_wallet": "Siubruoti pinigine slaptazodziu",
   "enter_forging_address": "Iveskite kalimo arba pool adresa",
   "error": "Klaida",
   "error_checking_status": "Klaida tikrinant busena",
@@ -269,10 +267,8 @@
   "import": "Importuoti",
   "import_mnemonic_description": "Iveskite 12 ar 24 zodziu atkurimo fraze pinigines atkurimui.",
   "import_to_wallet": "Importuoti i pinigine",
-  "import_wallet": "Importuoti pinigine",
   "import_wif_description": "Importuoti WIF privatu rakta kaip wpkh (Native SegWit) deskriptoriu i pinigine",
   "import_wif_wpkh": "Importuoti WIF rakta (wpkh)",
-  "importing": "Importuojama",
   "inbound": "Ieinantis",
   "incoming_payment": "Gautas mokejimas",
   "incorrect_word": "Neteisingas zodis",
@@ -541,7 +537,6 @@
   "passphrase": "Atkurimo fraze",
   "passphrase_mismatch": "Slaptazodziai nesutampa",
   "password": "Slaptazodis",
-  "password_mismatch": "Slaptazodziai nesutampa",
   "payment_confirmed": "Mokejimas patvirtintas",
   "payment_uri": "Mokejimo URI",
   "payment_uri_copied": "Mokejimo URI nukopijuotas i isvykimo sriti",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "senasis darbalaukis",
   "restore_branch_pocx": "mobilioji",
   "restore_history_found": "Lėšos rastos: {branches}",
-  "restore_no_history": "Šiai frazei ankstesnės istorijos nerasta — piniginė pradeda nuo nulio.",
   "reset": "Atstatyti",
   "reset_mining_config": "Atstatyti kasybos konfiguracija",
   "reset_mining_config_confirm": "Tai isvalys visus kasybos nustatymus ir gris i nustatymu vedli. Plot failai nebus istrinti. Ar tikrai?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Galite pasirinktinai siubruoti pinigine slaptazodziu. Tai apsaugo pinigines faila ir reikalauja slaptazodzio lesu prieigai.",
   "wallet_encryption_title": "Pinigines siubravimas",
   "wallet_encryption_warning": "Jei pamirssite si slaptazodi, vis tiek galite atkurti pinigine naudodami 24 zodziu fraze. Taciau teks sukurti nauja pinigine.",
-  "wallet_imported_success": "Pinigine '{name}' sekmingai importuota",
   "wallet_lock_status_all_locked": "Visos uzsifruotos pinigines uzrakintos",
   "wallet_lock_status_some_unlocked": "{unlocked} is {total} uzsifruotu pinigines atrakintos",
   "wallet_locked_message": "Si pinigine uzrakinta. Prasome ivesti slaptazodi laikinai atrakinti.",
   "wallet_name": "Pinigines pavadinimas",
-  "wallet_name_hint": "Pasirinkite pavadinima siai piniginei identifikuoti",
   "wallet_not_encrypted": "Nesiubruota",
-  "wallet_password": "Pinigines slaptazodis",
   "wallet_status_notifications": "Pinigines busenos pranesimai",
   "wallets": "Pinigine(s)",
   "warn": "Įspėj.",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "sinchronizuota prieš {seconds} s",
   "wallet_name_invalid_local": "Naudokite 1–32 raides, skaitmenis, „-“ arba „_“",
   "wallet_name_hint_local": "Tik raidės, skaitmenys, „-“ ir „_“",
-  "wallet_encryption_info_local": "Pasirinktinai užšifruokite piniginės sėklą diske slaptažodžiu. Jos prireiks piniginei atrakinti siunčiant. Vien atkūrimo frazė visada atkuria jūsų lėšas.",
-  "restore_probe_info": "Jūsų sėklos istorija tikrinama per Electrum importuojant — šakos su esamomis lėšomis aptinkamos automatiškai.",
   "psbt_join_unavailable_remote": "Sujungimas negalimas nuotoliniu režimu",
   "feature_unavailable_remote": "Ši funkcija negalima nuotoliniu (Electrum) režimu",
   "setup_solo_not_available_remote": "Solo kasybai reikia vietinio mazgo — negalima nuotoliniu (Electrum) režimu",
