+++ web-wallet/src/assets/locales/nl.json
@@ -140,7 +140,6 @@
   "confirm_send": "Bevestigen & Versturen",
   "confirm_send_message": "Controleer de transactiedetails voordat u verstuurt.",
   "confirm_transaction": "Transactie Bevestigen",
-  "confirm_wallet_password": "Bevestig Wachtwoord",
   "confirmations": "Bevestigingen",
   "confirmations_short": "bev.",
   "confirmed": "Bevestigd",
@@ -224,7 +223,6 @@
   "edit_contact": "Contact Bewerken",
   "effective_at_block": "Effectief bij blok",
   "effective_height": "Effectieve hoogte",
-  "encrypt_wallet": "Portemonnee versleutelen met wachtwoord",
   "enter_forging_address": "Voer forging- of pool-adres in",
   "error": "Fout",
   "error_checking_status": "Fout bij controleren status",
@@ -269,10 +267,8 @@
   "import": "Importeren",
   "import_mnemonic_description": "Voer uw 12 of 24 woorden herstelzin in om uw portemonnee te herstellen.",
   "import_to_wallet": "Importeren naar Portemonnee",
-  "import_wallet": "Portemonnee Importeren",
   "import_wif_description": "Importeer een WIF privésleutel als een wpkh (Native SegWit) descriptor naar portemonnee",
   "import_wif_wpkh": "WIF Sleutel Importeren (wpkh)",
-  "importing": "Importeren",
   "inbound": "Inkomend",
   "incoming_payment": "Inkomende Betaling",
   "incorrect_word": "Onjuist woord",
@@ -541,7 +537,6 @@
   "passphrase": "Herstelzin",
   "passphrase_mismatch": "Wachtwoordzinnen komen niet overeen",
   "password": "Wachtwoord",
-  "password_mismatch": "Wachtwoorden komen niet overeen",
   "payment_confirmed": "Betaling Bevestigd",
   "payment_uri": "Betaling URI",
   "payment_uri_copied": "Betaling URI gekopieerd naar klembord",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "oude desktop",
   "restore_branch_pocx": "mobiel",
   "restore_history_found": "Tegoed gevonden op: {branches}",
-  "restore_no_history": "Geen eerdere geschiedenis gevonden voor deze zin — de wallet begint blanco.",
   "reset": "Resetten",
   "reset_mining_config": "Mijnbouw Configuratie Resetten",
   "reset_mining_config_confirm": "Dit wist alle mijnbouw instellingen en keert terug naar de setup wizard. Plot bestanden worden niet verwijderd. Weet u het zeker?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "U kunt optioneel uw portemonnee versleutelen met een wachtwoord. Dit beschermt uw portemonnee-bestand en vereist het wachtwoord om toegang te krijgen tot uw fondsen.",
   "wallet_encryption_title": "Portemonnee Versleuteling",
   "wallet_encryption_warning": "Als u dit wachtwoord vergeet, kunt u uw portemonnee nog steeds herstellen met uw 24-woorden seed phrase. U moet echter een nieuwe portemonnee aanmaken.",
-  "wallet_imported_success": "Portemonnee '{name}' succesvol geïmporteerd",
   "wallet_lock_status_all_locked": "Alle versleutelde portemonnees vergrendeld",
   "wallet_lock_status_some_unlocked": "{unlocked} van {total} versleutelde portemonnees ontgrendeld",
   "wallet_locked_message": "Deze portemonnee is vergrendeld. Voer de wachtwoordzin in om tijdelijk te ontgrendelen.",
   "wallet_name": "Portemonnee Naam",
-  "wallet_name_hint": "Kies een naam om deze portemonnee te identificeren",
   "wallet_not_encrypted": "Niet versleuteld",
-  "wallet_password": "Portemonnee Wachtwoord",
   "wallet_status_notifications": "Portemonnee Status Meldingen",
   "wallets": "Portemonnee(s)",
   "warn": "Waarsch.",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "{seconds}s geleden gesynchroniseerd",
   "wallet_name_invalid_local": "Gebruik 1–32 letters, cijfers, '-' of '_'",
   "wallet_name_hint_local": "Alleen letters, cijfers, '-' en '_'",
-  "wallet_encryption_info_local": "Versleutel optioneel de wallet-seed op schijf met een wachtwoord. Je hebt het nodig om de wallet te ontgrendelen voor verzenden. De herstelzin alleen herstelt altijd je tegoed.",
-  "restore_probe_info": "De geschiedenis van je seed wordt tijdens het importeren via Electrum gecontroleerd — takken met bestaand tegoed worden automatisch gedetecteerd.",
   "psbt_join_unavailable_remote": "Samenvoegen is niet beschikbaar in externe modus",
   "feature_unavailable_remote": "Deze functie is niet beschikbaar in externe modus (Electrum)",
   "setup_solo_not_available_remote": "Solo-mining vereist een lokale node — niet beschikbaar in externe modus (Electrum)",
