+++ web-wallet/src/assets/locales/pl.json
@@ -140,7 +140,6 @@
   "confirm_send": "Potwierdz i wyslij",
   "confirm_send_message": "Sprawdz szczegoly transakcji przed wyslaniem.",
   "confirm_transaction": "Potwierdz transakcje",
-  "confirm_wallet_password": "Potwierdz haslo",
   "confirmations": "Potwierdzenia",
   "confirmations_short": "potw.",
   "confirmed": "Potwierdzone",
@@ -224,7 +223,6 @@
   "edit_contact": "Edytuj kontakt",
   "effective_at_block": "Skuteczne od bloku",
   "effective_height": "Wysokość obowiązywania",
-  "encrypt_wallet": "Zaszyfruj portfel haslem",
   "enter_forging_address": "Wprowadz adres kucia lub puli",
   "error": "Blad",
   "error_checking_status": "Blad sprawdzania statusu",
@@ -269,10 +267,8 @@
   "import": "Importuj",
   "import_mnemonic_description": "Wprowadz 12 lub 24 slowa frazy odzyskiwania aby przywrocic portfel.",
   "import_to_wallet": "Importuj do portfela",
-  "import_wallet": "Importuj portfel",
   "import_wif_description": "Importuj klucz prywatny WIF jako deskryptor wpkh (Native SegWit) do portfela",
   "import_wif_wpkh": "Importuj klucz WIF (wpkh)",
-  "importing": "Importowanie",
   "inbound": "Przychodzacy",
   "incoming_payment": "Platnosc przychodzaca",
   "incorrect_word": "Nieprawidlowe slowo",
@@ -541,7 +537,6 @@
   "passphrase": "Fraza odzyskiwania",
   "passphrase_mismatch": "Hasla sie nie zgadzaja",
   "password": "Haslo",
-  "password_mismatch": "Hasla sie nie zgadzaja",
   "payment_confirmed": "Platnosc potwierdzona",
   "payment_uri": "URI platnosci",
   "payment_uri_copied": "URI platnosci skopiowane do schowka",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "starszy desktop",
   "restore_branch_pocx": "mobilny",
   "restore_history_found": "Znaleziono środki na: {branches}",
-  "restore_no_history": "Nie znaleziono wcześniejszej historii dla tej frazy — portfel zaczyna od zera.",
   "reset": "Resetuj",
   "reset_mining_config": "Resetuj konfiguracje gornictwa",
   "reset_mining_config_confirm": "Spowoduje to wyczyszczenie wszystkich ustawien gornictwa i powrot do kreatora konfiguracji. Pliki plotow nie zostana usuniete. Czy na pewno?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Mozesz opcjonalnie zaszyfrowac portfel haslem. Chroni to plik portfela i wymaga hasla do uzyskania dostepu do srodkow.",
   "wallet_encryption_title": "Szyfrowanie portfela",
   "wallet_encryption_warning": "Jesli zapomnisz tego hasla, nadal mozesz odzyskac portfel uzywajac 24-slowowej frazy seed. Musisz jednak utworzyc nowy portfel.",
-  "wallet_imported_success": "Portfel '{name}' zaimportowany pomyslnie",
   "wallet_lock_status_all_locked": "Wszystkie zaszyfrowane portfele zablokowane",
   "wallet_lock_status_some_unlocked": "Odblokowano {unlocked} z {total} zaszyfrowanych portfeli",
   "wallet_locked_message": "Ten portfel jest zablokowany. Wprowadz haslo aby tymczasowo odblokowac.",
   "wallet_name": "Nazwa portfela",
-  "wallet_name_hint": "Wybierz nazwe aby zidentyfikowac ten portfel",
   "wallet_not_encrypted": "Niezaszyfrowany",
-  "wallet_password": "Haslo portfela",
   "wallet_status_notifications": "Powiadomienia o statusie portfela",
   "wallets": "Portfel(e)",
   "warn": "Ostrzeż.",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "zsynchronizowano {seconds} s temu",
   "wallet_name_invalid_local": "Użyj 1–32 liter, cyfr, „-” lub „_”",
   "wallet_name_hint_local": "Tylko litery, cyfry, „-” i „_”",
-  "wallet_encryption_info_local": "Opcjonalnie zaszyfruj ziarno portfela na dysku hasłem. Będzie potrzebne do odblokowania portfela w celu wysyłania. Sama fraza odzyskiwania zawsze przywraca Twoje środki.",
-  "restore_probe_info": "Historia Twojego ziarna jest sprawdzana przez Electrum podczas importu — gałęzie z istniejącymi środkami są wykrywane automatycznie.",
   "psbt_join_unavailable_remote": "Łączenie nie jest dostępne w trybie zdalnym",
   "feature_unavailable_remote": "Ta funkcja nie jest dostępna w trybie zdalnym (Electrum)",
   "setup_solo_not_available_remote": "Kopanie solo wymaga lokalnego węzła — niedostępne w trybie zdalnym (Electrum)",
