+++ web-wallet/src/assets/locales/pt-br.json
@@ -140,7 +140,6 @@
   "confirm_send": "Confirmar e enviar",
   "confirm_send_message": "Por favor, revise os detalhes da transação antes de enviar.",
   "confirm_transaction": "Confirmar transação",
-  "confirm_wallet_password": "Confirmar senha",
   "confirmations": "Confirmações",
   "confirmations_short": "conf",
   "confirmed": "Confirmado",
@@ -224,7 +223,6 @@
   "edit_contact": "Editar contato",
   "effective_at_block": "Efetivo no bloco",
   "effective_height": "Altura efetiva",
-  "encrypt_wallet": "Criptografar carteira com senha",
   "enter_forging_address": "Digite o endereço de forjamento ou de pool",
   "error": "Erro",
   "error_checking_status": "Erro ao verificar status",
@@ -269,10 +267,8 @@
   "import": "Importar",
   "import_mnemonic_description": "Digite sua frase de recuperação de 12 ou 24 palavras para restaurar sua carteira.",
   "import_to_wallet": "Importar para carteira",
-  "import_wallet": "Importar carteira",
   "import_wif_description": "Importar uma chave privada WIF como um descritor wpkh (Native SegWit) na carteira",
   "import_wif_wpkh": "Importar chave WIF (wpkh)",
-  "importing": "Importando",
   "inbound": "Entrada",
   "incoming_payment": "Pagamento recebido",
   "incorrect_word": "Palavra incorreta",
@@ -541,7 +537,6 @@
   "passphrase": "Frase de recuperação",
   "passphrase_mismatch": "As frases de senha não coincidem",
   "password": "Senha",
-  "password_mismatch": "As senhas não coincidem",
   "payment_confirmed": "Pagamento confirmado",
   "payment_uri": "URI de pagamento",
   "payment_uri_copied": "URI de pagamento copiada para a área de transferência",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "desktop antigo",
   "restore_branch_pocx": "móvel",
   "restore_history_found": "Fundos encontrados em: {branches}",
-  "restore_no_history": "Nenhum histórico anterior encontrado para esta frase — a carteira começa do zero.",
   "reset": "Redefinir",
   "reset_mining_config": "Redefinir configuração de mineração",
   "reset_mining_config_confirm": "Isso limpará todas as configurações de mineração e retornará ao assistente de configuração. Os arquivos de plot não serão excluídos. Tem certeza?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Opcionalmente, você pode criptografar sua carteira com uma senha. Isso protege o arquivo da sua carteira e requer a senha para acessar seus fundos.",
   "wallet_encryption_title": "Criptografia da carteira",
   "wallet_encryption_warning": "Se você esquecer esta senha, ainda poderá recuperar sua carteira usando sua frase semente de 24 palavras. No entanto, você precisará criar uma nova carteira.",
-  "wallet_imported_success": "Carteira '{name}' importada com sucesso",
   "wallet_lock_status_all_locked": "Todas as carteiras criptografadas bloqueadas",
   "wallet_lock_status_some_unlocked": "{unlocked} de {total} carteiras criptografadas desbloqueadas",
   "wallet_locked_message": "Esta carteira está bloqueada. Por favor, digite a frase de senha para desbloqueá-la temporariamente.",
   "wallet_name": "Nome da carteira",
-  "wallet_name_hint": "Escolha um nome para identificar esta carteira",
   "wallet_not_encrypted": "Não criptografada",
-  "wallet_password": "Senha da carteira",
   "wallet_status_notifications": "Notificações de status da carteira",
   "wallets": "Carteira(s)",
   "warn": "Aviso",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "sincronizado há {seconds}s",
   "wallet_name_invalid_local": "Use de 1 a 32 letras, dígitos, '-' ou '_'",
   "wallet_name_hint_local": "Apenas letras, dígitos, '-' e '_'",
-  "wallet_encryption_info_local": "Opcionalmente, criptografe a semente da carteira no disco com uma senha. Você precisará dela para desbloquear a carteira ao enviar. A frase de recuperação sozinha sempre restaura seus fundos.",
-  "restore_probe_info": "O histórico da sua semente é sondado via Electrum durante a importação — ramos com fundos existentes são detectados automaticamente.",
   "psbt_join_unavailable_remote": "Combinar não está disponível no modo remoto",
   "feature_unavailable_remote": "Este recurso não está disponível no modo remoto (Electrum)",
   "setup_solo_not_available_remote": "A mineração solo precisa de um nó local — não disponível no modo remoto (Electrum)",
