+++ web-wallet/src/assets/locales/ro.json
@@ -140,7 +140,6 @@
   "confirm_send": "Confirmă și trimite",
   "confirm_send_message": "Vă rugăm să verificați detaliile tranzacției înainte de a trimite.",
   "confirm_transaction": "Confirmă tranzacția",
-  "confirm_wallet_password": "Confirmă parola",
   "confirmations": "Confirmări",
   "confirmations_short": "conf",
   "confirmed": "Confirmat",
@@ -224,7 +223,6 @@
   "edit_contact": "Editează contact",
   "effective_at_block": "Efectiv la blocul",
   "effective_height": "Înălțimea efectivă",
-  "encrypt_wallet": "Criptează portofelul cu parolă",
   "enter_forging_address": "Introduceți adresa de forjare sau de pool",
   "error": "Eroare",
   "error_checking_status": "Eroare la verificarea stării",
@@ -269,10 +267,8 @@
   "import": "Importă",
   "import_mnemonic_description": "Introduceți fraza de recuperare de 12 sau 24 de cuvinte pentru a restaura portofelul.",
   "import_to_wallet": "Importă în portofel",
-  "import_wallet": "Importă portofel",
   "import_wif_description": "Importați o cheie privată WIF ca descriptor wpkh (Native SegWit) în portofel",
   "import_wif_wpkh": "Importă cheie WIF (wpkh)",
-  "importing": "Se importă",
   "inbound": "Intrare",
   "incoming_payment": "Plată primită",
   "incorrect_word": "Cuvânt incorect",
@@ -541,7 +537,6 @@
   "passphrase": "Frază de recuperare",
   "passphrase_mismatch": "Parolele nu se potrivesc",
   "password": "Parolă",
-  "password_mismatch": "Parolele nu se potrivesc",
   "payment_confirmed": "Plată confirmată",
   "payment_uri": "URI plată",
   "payment_uri_copied": "URI-ul de plată a fost copiat în clipboard",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "desktop vechi",
   "restore_branch_pocx": "mobil",
   "restore_history_found": "Fonduri găsite pe: {branches}",
-  "restore_no_history": "Nu s-a găsit istoric anterior pentru această frază — portofelul începe de la zero.",
   "reset": "Resetare",
   "reset_mining_config": "Resetează configurația mineritului",
   "reset_mining_config_confirm": "Aceasta va șterge toate setările de minerit și va reveni la asistentul de configurare. Fișierele plot nu vor fi șterse. Sunteți sigur?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Puteți cripta opțional portofelul cu o parolă. Aceasta protejează fișierul portofelului și necesită parola pentru a accesa fondurile.",
   "wallet_encryption_title": "Criptare portofel",
   "wallet_encryption_warning": "Dacă uitați această parolă, puteți încă recupera portofelul folosind fraza de 24 de cuvinte. Cu toate acestea, va trebui să creați un portofel nou.",
-  "wallet_imported_success": "Portofelul '{name}' a fost importat cu succes",
   "wallet_lock_status_all_locked": "Toate portofelele criptate blocate",
   "wallet_lock_status_some_unlocked": "Deblocate {unlocked} din {total} portofele criptate",
   "wallet_locked_message": "Acest portofel este blocat. Vă rugăm să introduceți parola pentru a-l debloca temporar.",
   "wallet_name": "Nume portofel",
-  "wallet_name_hint": "Alegeți un nume pentru a identifica acest portofel",
   "wallet_not_encrypted": "Necriptat",
-  "wallet_password": "Parolă portofel",
   "wallet_status_notifications": "Notificări stare portofel",
   "wallets": "Portofel(e)",
   "warn": "Avert.",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "sincronizat acum {seconds}s",
   "wallet_name_invalid_local": "Folosește 1–32 litere, cifre, '-' sau '_'",
   "wallet_name_hint_local": "Doar litere, cifre, '-' și '_'",
-  "wallet_encryption_info_local": "Opțional, criptează seed-ul portofelului pe disc cu o parolă. Îți va fi necesară pentru a debloca portofelul la trimitere. Fraza de recuperare singură îți restaurează întotdeauna fondurile.",
-  "restore_probe_info": "Istoricul seed-ului tău este sondat prin Electrum în timpul importului — ramurile cu fonduri existente sunt detectate automat.",
   "psbt_join_unavailable_remote": "Combinarea nu este disponibilă în modul la distanță",
   "feature_unavailable_remote": "Această funcție nu este disponibilă în modul la distanță (Electrum)",
   "setup_solo_not_available_remote": "Mineritul solo necesită un nod local — indisponibil în modul la distanță (Electrum)",
