+++ web-wallet/src/assets/locales/ru.json
@@ -140,7 +140,6 @@
   "confirm_send": "Подтвердить и отправить",
   "confirm_send_message": "Пожалуйста, проверьте детали транзакции перед отправкой.",
   "confirm_transaction": "Подтвердить транзакцию",
-  "confirm_wallet_password": "Подтвердите пароль",
   "confirmations": "Подтверждений",
   "confirmations_short": "подтв.",
   "confirmed": "Подтверждено",
@@ -224,7 +223,6 @@
   "edit_contact": "Редактировать контакт",
   "effective_at_block": "Вступает в силу на блоке",
   "effective_height": "Блок вступления в силу",
-  "encrypt_wallet": "Зашифровать кошелек паролем",
   "enter_forging_address": "Введите адрес форжинга или пула",
   "error": "Ошибка",
   "error_checking_status": "Ошибка проверки статуса",
@@ -269,10 +267,8 @@
   "import": "Импорт",
   "import_mnemonic_description": "Введите вашу фразу восстановления из 12 или 24 слов для восстановления кошелька.",
   "import_to_wallet": "Импортировать в кошелек",
-  "import_wallet": "Импортировать кошелек",
   "import_wif_description": "Импортировать WIF приватный ключ как wpkh (Native SegWit) дескриптор в кошелек",
   "import_wif_wpkh": "Импорт WIF ключа (wpkh)",
-  "importing": "Импорт",
   "inbound": "Входящий",
   "incoming_payment": "Входящий платеж",
   "incorrect_word": "Неправильное слово",
@@ -541,7 +537,6 @@
   "passphrase": "Фраза восстановления",
   "passphrase_mismatch": "Парольные фразы не совпадают",
   "password": "Пароль",
-  "password_mismatch": "Пароли не совпадают",
   "payment_confirmed": "Платеж подтвержден",
   "payment_uri": "URI платежа",
   "payment_uri_copied": "URI платежа скопирован в буфер обмена",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "старый десктоп",
   "restore_branch_pocx": "мобильный",
   "restore_history_found": "Средства найдены на: {branches}",
-  "restore_no_history": "Для этой фразы не найдено прежней истории — кошелёк начинает с чистого листа.",
   "reset": "Сброс",
   "reset_mining_config": "Сбросить настройки майнинга",
   "reset_mining_config_confirm": "Это очистит все настройки майнинга и вернёт к мастеру настройки. Файлы плотов не будут удалены. Вы уверены?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Вы можете опционально зашифровать свой кошелек паролем. Это защищает файл кошелька и требует пароль для доступа к вашим средствам.",
   "wallet_encryption_title": "Шифрование кошелька",
   "wallet_encryption_warning": "Если вы забудете этот пароль, вы всё ещё сможете восстановить кошелек, используя вашу фразу из 24 слов. Однако вам придется создать новый кошелек.",
-  "wallet_imported_success": "Кошелек '{name}' успешно импортирован",
   "wallet_lock_status_all_locked": "Все зашифрованные кошельки заблокированы",
   "wallet_lock_status_some_unlocked": "Разблокировано {unlocked} из {total} зашифрованных кошельков",
   "wallet_locked_message": "Этот кошелек заблокирован. Пожалуйста, введите пароль для временной разблокировки.",
   "wallet_name": "Имя кошелька",
-  "wallet_name_hint": "Выберите имя для идентификации этого кошелька",
   "wallet_not_encrypted": "Не зашифрован",
-  "wallet_password": "Пароль кошелька",
   "wallet_status_notifications": "Уведомления о статусе кошелька",
   "wallets": "Кошелек(и)",
   "warn": "Предупр.",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "синхронизировано {seconds} с назад",
   "wallet_name_invalid_local": "Используйте 1–32 буквы, цифры, «-» или «_»",
   "wallet_name_hint_local": "Только буквы, цифры, «-» и «_»",
-  "wallet_encryption_info_local": "При желании зашифруйте сид кошелька на диске паролем. Он понадобится для разблокировки кошелька при отправке. Одна лишь фраза восстановления всегда вернёт ваши средства.",
-  "restore_probe_info": "История вашего сида проверяется через Electrum во время импорта — ветви с существующими средствами обнаруживаются автоматически.",
   "psbt_join_unavailable_remote": "Объединение недоступно в удалённом режиме",
   "feature_unavailable_remote": "Эта функция недоступна в удалённом режиме (Electrum)",
   "setup_solo_not_available_remote": "Соло-майнингу нужен локальный узел — недоступно в удалённом режиме (Electrum)",
