+++ web-wallet/src/assets/locales/sk.json
@@ -140,7 +140,6 @@
   "confirm_send": "Potvrdit a odoslat",
   "confirm_send_message": "Skontrolujte podrobnosti transakcie pred odoslanim.",
   "confirm_transaction": "Potvrdit transakciu",
-  "confirm_wallet_password": "Potvrdit heslo",
   "confirmations": "Potvrdenia",
   "confirmations_short": "potv.",
   "confirmed": "Potvrdene",
@@ -224,7 +223,6 @@
   "edit_contact": "Upravit kontakt",
   "effective_at_block": "Ucinne od bloku",
   "effective_height": "Účinná výška",
-  "encrypt_wallet": "Zasifrovat penazenku heslom",
   "enter_forging_address": "Zadajte adresu kovania alebo poolu",
   "error": "Chyba",
   "error_checking_status": "Chyba pri kontrole stavu",
@@ -269,10 +267,8 @@
   "import": "Importovat",
   "import_mnemonic_description": "Zadajte svoju 12 alebo 24 slovnu obnovovaciu frazu pre obnovenie penazenky.",
   "import_to_wallet": "Importovat do penazenky",
-  "import_wallet": "Importovat penazenku",
   "import_wif_description": "Importovat WIF sukromny kluc ako wpkh (Native SegWit) deskriptor do penazenky",
   "import_wif_wpkh": "Importovat WIF kluc (wpkh)",
-  "importing": "Importovanie",
   "inbound": "Prichadzajuci",
   "incoming_payment": "Prichadzajuca platba",
   "incorrect_word": "Nespravne slovo",
@@ -541,7 +537,6 @@
   "passphrase": "Obnovovacia fraza",
   "passphrase_mismatch": "Hesla sa nezhoduju",
   "password": "Heslo",
-  "password_mismatch": "Hesla sa nezhoduju",
   "payment_confirmed": "Platba potvrdena",
   "payment_uri": "URI platby",
   "payment_uri_copied": "URI platby skopirovane do schranky",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "starší desktop",
   "restore_branch_pocx": "mobilný",
   "restore_history_found": "Prostriedky nájdené na: {branches}",
-  "restore_no_history": "Pre túto frázu sa nenašla žiadna skoršia história — peňaženka začína odznova.",
   "reset": "Resetovat",
   "reset_mining_config": "Resetovat konfiguraciu tazby",
   "reset_mining_config_confirm": "Tym sa vymazu vsetky nastavenia tazby a vratite sa do sprievodcu nastavenim. Subory plotov nebudu vymazane. Naozaj?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Mozete volitelne zasifrovat svoju penazenku heslom. To chrani subor penazenky a vyzaduje heslo pre pristup k prostriedkom.",
   "wallet_encryption_title": "Sifrovanie penazenky",
   "wallet_encryption_warning": "Ak toto heslo zabudnete, mozete stale obnovit penazenku pomocou svojej 24slovnej seed frazy. Budete vsak musiet vytvorit novu penazenku.",
-  "wallet_imported_success": "Penazenka '{name}' uspesne importovana",
   "wallet_lock_status_all_locked": "Vsetky sifrovane penazenky uzamknute",
   "wallet_lock_status_some_unlocked": "Odomknute {unlocked} z {total} sifrovanych penazeniek",
   "wallet_locked_message": "Tato penazenka je uzamknuta. Zadajte heslo pre docasne odomknutie.",
   "wallet_name": "Nazov penazenky",
-  "wallet_name_hint": "Zvolte nazov pre identifikaciu tejto penazenky",
   "wallet_not_encrypted": "Nezasifrovane",
-  "wallet_password": "Heslo penazenky",
   "wallet_status_notifications": "Oznamenia o stave penazenky",
   "wallets": "Penazenka(y)",
   "warn": "Varov.",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "synchronizované pred {seconds} s",
   "wallet_name_invalid_local": "Použite 1–32 písmen, číslic, „-“ alebo „_“",
   "wallet_name_hint_local": "Iba písmená, číslice, „-“ a „_“",
-  "wallet_encryption_info_local": "Voliteľne zašifrujte semeno peňaženky na disku heslom. Budete ho potrebovať na odomknutie peňaženky pri odosielaní. Samotná obnovovacia fráza vždy obnoví vaše prostriedky.",
-  "restore_probe_info": "História vášho semena sa počas importu preveruje cez Electrum — vetvy s existujúcimi prostriedkami sa zisťujú automaticky.",
   "psbt_join_unavailable_remote": "Zlúčenie nie je vo vzdialenom režime k dispozícii",
   "feature_unavailable_remote": "Táto funkcia nie je k dispozícii vo vzdialenom (Electrum) režime",
   "setup_solo_not_available_remote": "Sólo ťažba vyžaduje lokálny uzol — nie je k dispozícii vo vzdialenom (Electrum) režime",
