+++ web-wallet/src/assets/locales/sr.json
@@ -140,7 +140,6 @@
   "confirm_send": "Потврди и пошаљи",
   "confirm_send_message": "Молимо прегледајте детаље трансакције пре слања.",
   "confirm_transaction": "Потврдите трансакцију",
-  "confirm_wallet_password": "Потврдите лозинку",
   "confirmations": "Потврде",
   "confirmations_short": "потв.",
   "confirmed": "Потврђено",
@@ -224,7 +223,6 @@
   "edit_contact": "Уреди контакт",
   "effective_at_block": "На снази од блока",
   "effective_height": "Ефективна висина",
-  "encrypt_wallet": "Шифруј новчаник лозинком",
   "enter_forging_address": "Унесите адресу ковања или пула",
   "error": "Грешка",
   "error_checking_status": "Грешка при провери статуса",
@@ -269,10 +267,8 @@
   "import": "Увези",
   "import_mnemonic_description": "Унесите вашу фразу за опоравак од 12 или 24 речи да опоравите новчаник.",
   "import_to_wallet": "Увези у новчаник",
-  "import_wallet": "Увези новчаник",
   "import_wif_description": "Увезите WIF приватни кључ као wpkh (Native SegWit) дескриптор у новчаник",
   "import_wif_wpkh": "Увези WIF кључ (wpkh)",
-  "importing": "Увожење",
   "inbound": "Долазни",
   "incoming_payment": "Долазна уплата",
   "incorrect_word": "Погрешна реч",
@@ -541,7 +537,6 @@
   "passphrase": "Фраза за опоравак",
   "passphrase_mismatch": "Лозинке се не подударају",
   "password": "Лозинка",
-  "password_mismatch": "Лозинке се не подударају",
   "payment_confirmed": "Плаћање потврђено",
   "payment_uri": "URI плаћања",
   "payment_uri_copied": "URI плаћања је копиран у привремену меморију",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "стари десктоп",
   "restore_branch_pocx": "мобилни",
   "restore_history_found": "Средства пронађена на: {branches}",
-  "restore_no_history": "За ову фразу није пронађена ранија историја — новчаник почиње испочетка.",
   "reset": "Ресетуј",
   "reset_mining_config": "Ресетуј конфигурацију рударења",
   "reset_mining_config_confirm": "Ово ће обрисати сва подешавања рударења и вратити чаробњак за подешавање. Плот фајлови неће бити обрисани. Да ли сте сигурни?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Можете опционо шифровати новчаник лозинком. Ово штити датотеку новчаника и захтева лозинку за приступ вашим средствима.",
   "wallet_encryption_title": "Шифровање новчаника",
   "wallet_encryption_warning": "Ако заборавите ову лозинку, и даље можете опоравити новчаник користећи фразу од 24 речи. Међутим, мораћете креирати нови новчаник.",
-  "wallet_imported_success": "Новчаник '{name}' је успешно увезен",
   "wallet_lock_status_all_locked": "Сви шифровани новчаници закључани",
   "wallet_lock_status_some_unlocked": "Откључано {unlocked} од {total} шифрованих новчаника",
   "wallet_locked_message": "Овај новчаник је закључан. Молимо унесите лозинку за привремено откључавање.",
   "wallet_name": "Име новчаника",
-  "wallet_name_hint": "Изаберите име за идентификацију овог новчаника",
   "wallet_not_encrypted": "Није шифровано",
-  "wallet_password": "Лозинка новчаника",
   "wallet_status_notifications": "Обавештења о статусу новчаника",
   "wallets": "Новчаник(ци)",
   "warn": "Упоз.",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "синхронизовано пре {seconds} с",
   "wallet_name_invalid_local": "Користите 1–32 слова, цифре, „-“ или „_“",
   "wallet_name_hint_local": "Само слова, цифре, „-“ и „_“",
-  "wallet_encryption_info_local": "Опционо шифрујте семе новчаника на диску лозинком. Биће вам потребна да откључате новчаник за слање. Сама фраза за опоравак увек враћа ваша средства.",
-  "restore_probe_info": "Историја вашег семена се испитује преко Electrum-а током увоза — гране са постојећим средствима се аутоматски откривају.",
   "psbt_join_unavailable_remote": "Спајање није доступно у удаљеном режиму",
   "feature_unavailable_remote": "Ова функција није доступна у удаљеном (Electrum) режиму",
   "setup_solo_not_available_remote": "Соло рударење захтева локални чвор — није доступно у удаљеном (Electrum) режиму",
