+++ web-wallet/src/assets/locales/tr.json
@@ -140,7 +140,6 @@
   "confirm_send": "Onayla ve Gönder",
   "confirm_send_message": "Göndermeden önce işlem detaylarını inceleyin.",
   "confirm_transaction": "İşlemi Onayla",
-  "confirm_wallet_password": "Şifreyi Onayla",
   "confirmations": "Onaylar",
   "confirmations_short": "onay",
   "confirmed": "Onaylandı",
@@ -224,7 +223,6 @@
   "edit_contact": "Kişiyi Düzenle",
   "effective_at_block": "Geçerli olduğu blok",
   "effective_height": "Yürürlük yüksekliği",
-  "encrypt_wallet": "Cüzdanı şifreyle şifrele",
   "enter_forging_address": "Dövme veya havuz adresi girin",
   "error": "Hata",
   "error_checking_status": "Durum kontrol edilirken hata",
@@ -269,10 +267,8 @@
   "import": "İçe Aktar",
   "import_mnemonic_description": "Cüzdanınızı geri yüklemek için 12 veya 24 kelimelik kurtarma ifadenizi girin.",
   "import_to_wallet": "Cüzdana İçe Aktar",
-  "import_wallet": "Cüzdan İçe Aktar",
   "import_wif_description": "Bir WIF özel anahtarını cüzdana wpkh (Native SegWit) tanımlayıcısı olarak içe aktarın",
   "import_wif_wpkh": "WIF Anahtarı İçe Aktar (wpkh)",
-  "importing": "İçe Aktarılıyor",
   "inbound": "Gelen",
   "incoming_payment": "Gelen Ödeme",
   "incorrect_word": "Yanlış kelime",
@@ -541,7 +537,6 @@
   "passphrase": "Kurtarma İfadesi",
   "passphrase_mismatch": "Parolalar eşleşmiyor",
   "password": "Şifre",
-  "password_mismatch": "Şifreler eşleşmiyor",
   "payment_confirmed": "Ödeme Onaylandı",
   "payment_uri": "Ödeme URI",
   "payment_uri_copied": "Ödeme URI panoya kopyalandı",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "eski masaüstü",
   "restore_branch_pocx": "mobil",
   "restore_history_found": "Şurada fon bulundu: {branches}",
-  "restore_no_history": "Bu ifade için önceki bir geçmiş bulunamadı — cüzdan sıfırdan başlıyor.",
   "reset": "Sıfırla",
   "reset_mining_config": "Madencilik Yapılandırmasını Sıfırla",
   "reset_mining_config_confirm": "Bu, tüm madencilik ayarlarını temizleyecek ve kurulum sihirbazına dönecektir. Plot dosyaları silinmeyecek. Emin misiniz?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Cüzdanınızı isteğe bağlı olarak şifreyle şifreleyebilirsiniz. Bu, cüzdan dosyanızı korur ve fonlarınıza erişmek için şifre gerektirir.",
   "wallet_encryption_title": "Cüzdan Şifreleme",
   "wallet_encryption_warning": "Bu şifreyi unutursanız, yine de 24 kelimelik tohum ifadenizi kullanarak cüzdanınızı kurtarabilirsiniz. Ancak, yeni bir cüzdan oluşturmanız gerekecek.",
-  "wallet_imported_success": "'{name}' cüzdanı başarıyla içe aktarıldı",
   "wallet_lock_status_all_locked": "Tüm şifreli cüzdanlar kilitli",
   "wallet_lock_status_some_unlocked": "{total} şifreli cüzdandan {unlocked} tanesi kilitsiz",
   "wallet_locked_message": "Bu cüzdan kilitli. Geçici olarak açmak için parolayı girin.",
   "wallet_name": "Cüzdan Adı",
-  "wallet_name_hint": "Bu cüzdanı tanımlamak için bir isim seçin",
   "wallet_not_encrypted": "Şifrelenmemiş",
-  "wallet_password": "Cüzdan Şifresi",
   "wallet_status_notifications": "Cüzdan Durum Bildirimleri",
   "wallets": "Cüzdan(lar)",
   "warn": "Uyarı",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "{seconds} sn önce eşitlendi",
   "wallet_name_invalid_local": "1–32 harf, rakam, '-' veya '_' kullanın",
   "wallet_name_hint_local": "Yalnızca harf, rakam, '-' ve '_'",
-  "wallet_encryption_info_local": "İsteğe bağlı olarak cüzdan tohumunu diskte bir parolayla şifreleyin. Gönderim için cüzdanın kilidini açmak üzere buna ihtiyacınız olacak. Yalnızca kurtarma ifadesi her zaman paranızı geri yükler.",
-  "restore_probe_info": "Tohumunuzun geçmişi içe aktarma sırasında Electrum üzerinden araştırılır — mevcut fon içeren dallar otomatik olarak algılanır.",
   "psbt_join_unavailable_remote": "Birleştirme uzak modda kullanılamaz",
   "feature_unavailable_remote": "Bu özellik uzak (Electrum) modda kullanılamaz",
   "setup_solo_not_available_remote": "Solo madencilik yerel bir düğüm gerektirir — uzak (Electrum) modda kullanılamaz",
