+++ web-wallet/src/assets/locales/uk.json
@@ -140,7 +140,6 @@
   "confirm_send": "Підтвердити та надіслати",
   "confirm_send_message": "Будь ласка, перевірте деталі транзакції перед надсиланням.",
   "confirm_transaction": "Підтвердити транзакцію",
-  "confirm_wallet_password": "Підтвердіть пароль",
   "confirmations": "Підтвердження",
   "confirmations_short": "підтв.",
   "confirmed": "Підтверджено",
@@ -224,7 +223,6 @@
   "edit_contact": "Редагувати контакт",
   "effective_at_block": "Набуває чинності на блоці",
   "effective_height": "Висота набуття чинності",
-  "encrypt_wallet": "Зашифрувати гаманець паролем",
   "enter_forging_address": "Введіть адресу ковки або пулу",
   "error": "Помилка",
   "error_checking_status": "Помилка перевірки статусу",
@@ -269,10 +267,8 @@
   "import": "Імпортувати",
   "import_mnemonic_description": "Введіть вашу фразу відновлення з 12 або 24 слів для відновлення гаманця.",
   "import_to_wallet": "Імпортувати до гаманця",
-  "import_wallet": "Імпортувати гаманець",
   "import_wif_description": "Імпортувати приватний ключ WIF як дескриптор wpkh (Native SegWit) до гаманця",
   "import_wif_wpkh": "Імпортувати WIF ключ (wpkh)",
-  "importing": "Імпорт",
   "inbound": "Вхідний",
   "incoming_payment": "Вхідний платіж",
   "incorrect_word": "Неправильне слово",
@@ -541,7 +537,6 @@
   "passphrase": "Фраза відновлення",
   "passphrase_mismatch": "Парольні фрази не збігаються",
   "password": "Пароль",
-  "password_mismatch": "Паролі не збігаються",
   "payment_confirmed": "Платіж підтверджено",
   "payment_uri": "URI платежу",
   "payment_uri_copied": "URI платежу скопійовано в буфер обміну",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "старий десктоп",
   "restore_branch_pocx": "мобільний",
   "restore_history_found": "Кошти знайдено на: {branches}",
-  "restore_no_history": "Для цієї фрази не знайдено попередньої історії — гаманець починає з нуля.",
   "reset": "Скинути",
   "reset_mining_config": "Скинути конфігурацію майнінгу",
   "reset_mining_config_confirm": "Це очистить усі налаштування майнінгу та поверне до майстра налаштування. Файли плотів не буде видалено. Ви впевнені?",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "Ви можете опціонально зашифрувати свій гаманець паролем. Це захищає файл гаманця та вимагає пароль для доступу до ваших коштів.",
   "wallet_encryption_title": "Шифрування гаманця",
   "wallet_encryption_warning": "Якщо ви забудете цей пароль, ви все ще зможете відновити гаманець, використовуючи вашу фразу з 24 слів. Однак вам доведеться створити новий гаманець.",
-  "wallet_imported_success": "Гаманець '{name}' успішно імпортовано",
   "wallet_lock_status_all_locked": "Усі зашифровані гаманці заблоковано",
   "wallet_lock_status_some_unlocked": "Розблоковано {unlocked} з {total} зашифрованих гаманців",
   "wallet_locked_message": "Цей гаманець заблоковано. Будь ласка, введіть пароль для тимчасового розблокування.",
   "wallet_name": "Назва гаманця",
-  "wallet_name_hint": "Виберіть назву для ідентифікації цього гаманця",
   "wallet_not_encrypted": "Не зашифровано",
-  "wallet_password": "Пароль гаманця",
   "wallet_status_notifications": "Сповіщення про статус гаманця",
   "wallets": "Гаманець(ці)",
   "warn": "Попер.",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "синхронізовано {seconds} с тому",
   "wallet_name_invalid_local": "Використовуйте 1–32 літери, цифри, «-» або «_»",
   "wallet_name_hint_local": "Лише літери, цифри, «-» та «_»",
-  "wallet_encryption_info_local": "За бажанням зашифруйте насіння гаманця на диску паролем. Він знадобиться для розблокування гаманця для надсилання. Сама лише фраза відновлення завжди відновлює ваші кошти.",
-  "restore_probe_info": "Історія вашого насіння перевіряється через Electrum під час імпортування — гілки з наявними коштами виявляються автоматично.",
   "psbt_join_unavailable_remote": "Об'єднання недоступне у віддаленому режимі",
   "feature_unavailable_remote": "Ця функція недоступна у віддаленому режимі (Electrum)",
   "setup_solo_not_available_remote": "Соло-майнінгу потрібен локальний вузол — недоступно у віддаленому режимі (Electrum)",
