+++ web-wallet/src/assets/locales/zh-cn.json
@@ -140,7 +140,6 @@
   "confirm_send": "确认并发送",
   "confirm_send_message": "发送前请检查交易详情。",
   "confirm_transaction": "确认交易",
-  "confirm_wallet_password": "确认密码",
   "confirmations": "确认数",
   "confirmations_short": "确认",
   "confirmed": "已确认",
@@ -224,7 +223,6 @@
   "edit_contact": "编辑联系人",
   "effective_at_block": "生效于区块",
   "effective_height": "生效高度",
-  "encrypt_wallet": "使用密码加密钱包",
   "enter_forging_address": "输入锻造或矿池地址",
   "error": "错误",
   "error_checking_status": "检查状态时出错",
@@ -269,10 +267,8 @@
   "import": "导入",
   "import_mnemonic_description": "输入您的12或24个助记词以恢复钱包。",
   "import_to_wallet": "导入到钱包",
-  "import_wallet": "导入钱包",
   "import_wif_description": "将 WIF 私钥作为 wpkh（原生隔离见证）描述符导入到钱包",
   "import_wif_wpkh": "导入 WIF 密钥 (wpkh)",
-  "importing": "导入中",
   "inbound": "入站",
   "incoming_payment": "收款",
   "incorrect_word": "单词错误",
@@ -541,7 +537,6 @@
   "passphrase": "恢复短语",
   "passphrase_mismatch": "密码短语不匹配",
   "password": "密码",
-  "password_mismatch": "密码不匹配",
   "payment_confirmed": "付款已确认",
   "payment_uri": "付款URI",
   "payment_uri_copied": "付款URI已复制到剪贴板",
@@ -590,7 +585,6 @@
   "restore_branch_legacy": "旧版桌面",
   "restore_branch_pocx": "移动端",
   "restore_history_found": "在以下分支发现资金：{branches}",
-  "restore_no_history": "未找到此助记词的既往历史记录 — 钱包将全新开始。",
   "reset": "重置",
   "reset_mining_config": "重置挖矿配置",
   "reset_mining_config_confirm": "这将清除所有挖矿设置并返回设置向导。绘图文件不会被删除。确定要继续吗？",
@@ -843,14 +837,11 @@
   "wallet_encryption_info": "您可以选择使用密码加密钱包。这将保护您的钱包文件，访问资金时需要密码。",
   "wallet_encryption_title": "钱包加密",
   "wallet_encryption_warning": "如果忘记此密码，您仍可使用24个助记词恢复钱包。但是，您需要创建新钱包。",
-  "wallet_imported_success": "钱包 '{name}' 导入成功",
   "wallet_lock_status_all_locked": "所有加密钱包已锁定",
   "wallet_lock_status_some_unlocked": "{total} 个加密钱包中的 {unlocked} 个已解锁",
   "wallet_locked_message": "此钱包已锁定。请输入密码以临时解锁。",
   "wallet_name": "钱包名称",
-  "wallet_name_hint": "选择一个名称来标识此钱包",
   "wallet_not_encrypted": "未加密",
-  "wallet_password": "钱包密码",
   "wallet_status_notifications": "钱包状态通知",
   "wallets": "钱包",
   "warn": "警告",
@@ -1204,8 +1195,6 @@
   "electrum_synced_ago": "{seconds} 秒前已同步",
   "wallet_name_invalid_local": "使用 1–32 个字母、数字、“-”或“_”",
   "wallet_name_hint_local": "仅限字母、数字、“-”和“_”",
-  "wallet_encryption_info_local": "可选择使用密码在磁盘上加密钱包种子。发送时您需要用它来解锁钱包。仅凭恢复短语始终可以恢复您的资金。",
-  "restore_probe_info": "导入期间，系统会通过 Electrum 探测您种子的历史记录——含有资金的分支会被自动检测到。",
   "psbt_join_unavailable_remote": "远程模式下无法使用合并",
   "feature_unavailable_remote": "此功能在远程（Electrum）模式下不可用",
   "setup_solo_not_available_remote": "独立挖矿需要本地节点——在远程（Electrum）模式下不可用",
